% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.DatasetExperiment}
\alias{as.DatasetExperiment}
\title{Convert a SummarizedExperiment to DatasetExperiment}
\usage{
as.DatasetExperiment(obj)
}
\arguments{
\item{obj}{a SummarizedExperiment object}
}
\value{
a DatasetExperiment object
}
\description{
Converts a SummarizedExperiment to DatasetExperiment. The assay data is
transposed, and colData and rowData switched to match. struct specific
slots such as "name" and "description" are extracted from the metaData.
}
