% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetExperiment_class.R
\name{as.SummarizedExperiment,DatasetExperiment-method}
\alias{as.SummarizedExperiment,DatasetExperiment-method}
\title{Convert a DatasetExperiment to SummarizedExperiment}
\usage{
\S4method{as.SummarizedExperiment}{DatasetExperiment}(obj)
}
\arguments{
\item{obj}{a DatasetExperiment object}
}
\value{
a SummarizedExperiment object
}
\description{
Converts a DatasetExperiment to SummarizedExperiment. The assay data is
transposed, and colData and rowData switched to match. struct specific
slots such as "name" and "description" are stored in the metaData.
}
