% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_data_frame}
\alias{as_data_frame}
\title{convert to data.frame}
\usage{
as_data_frame(M)
}
\arguments{
\item{M}{a struct object}

\item{...}{other inputs passed through this function}
}
\value{
a data.frame containing outputs from an object
}
\description{
Most often used with univariate statistics to gather all the different outputs in a consistent format.
}
