% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as_layer}
\alias{as_layer}
\title{Convert layer entity to ggplot2 layer}
\usage{
as_layer(obj, ...)
}
\arguments{
\item{obj}{a layer_entity object}

\item{...}{additional arguments}
}
\value{
a ggplot2 layer object
}
\description{
Convert layer entity to ggplot2 layer
}
