% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{available_presets}
\alias{available_presets}
\title{Get available presets for a chart}
\usage{
available_presets(obj, name)
}
\arguments{
\item{obj}{a chart object}

\item{name}{character string naming the slot to check}
}
\value{
character vector of available preset names
}
\description{
Get available presets for a chart
}
