% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_term_class.R
\name{c,ontology_list-method}
\alias{c,ontology_list-method}
\title{catenate ontology_lists}
\usage{
\S4method{c}{ontology_list}(x, ...)
}
\arguments{
\item{x}{an ontology_list()}

\item{...}{any number of ontology_list() or ontology_item() objects to catenate}
}
\value{
an ontology_list()
}
\description{
ontology_lists can be catenated with other ontology lists or with ontology_items.
}
