% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_term_class.R
\name{$,ontology_list-method}
\alias{$,ontology_list-method}
\title{Get/set ontology_list slots}
\usage{
\S4method{$}{ontology_list}(x, name)
}
\arguments{
\item{x}{An ontology_term object}

\item{name}{The name of the slot to access}
}
\value{
Slot value
}
\description{
Dollar syntax can be used to as a shortcut for getting
values for ontology_list objects.
}
\examples{
\dontrun{
OL = ontology_list('STATO:0000046')
OL$terms
}
}
