% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_term_class.R
\name{$,ontology_term-method}
\alias{$,ontology_term-method}
\title{Get/set ontology term slots}
\usage{
\S4method{$}{ontology_term}(x, name)
}
\arguments{
\item{x}{An ontology_term object}

\item{name}{The name of the slot to access}
}
\value{
Slot value
}
\description{
Dollar syntax can be used to as a shortcut for getting
values for ontology_term objects.
}
\examples{
\dontrun{
OT = ontology_term(id='STATO:0000046')
OT$ontology # returns "STATO"
}
}
