% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.ids}
\alias{chart.ids}
\title{valid chart names for an object}
\usage{
chart.ids(obj)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of chart to plot}
}
\description{
valid chart names for an object
}
\seealso{
Other chart functions: \code{\link{chart.name}},
  \code{\link{chart.obj}}, \code{\link{chart.opt}},
  \code{\link{chart.plot}}, \code{\link{is.chart}}
}
