% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.name}
\alias{chart.name}
\title{chart names}
\usage{
chart.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
list of chart names
}
\description{
return a list of valid parameter names for an object
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.obj}}, \code{\link{chart.opt}},
  \code{\link{chart.plot}}, \code{\link{is.chart}}
}
