% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.names}
\alias{chart.names}
\title{chart names}
\usage{
chart.names(obj, ...)
}
\arguments{
\item{obj}{a object from the *struct* package}

\item{...}{optional inputs (see details)}
}
\value{
list of chart names
}
\description{
Returns a list of valid charts for a struct object
}
\details{
The chart.names method searches chart objects that specify the input object
type as an input.

Optional input \code{ret = 'char'} is the default and will return a character
vector of object names. \code{ret = 'obj'} will return a list containing the
named chart objects.
}
\examples{
M = example_model()
chart.names(M) # 'example_chart'
chart.names(M,'char') # as above
chart.names(M,'obj') # returns a list of chart objects

}
