% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.obj}
\alias{chart.obj}
\alias{chart.obj<-}
\title{get/set chart object}
\usage{
chart.obj(obj, name)

chart.obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of chart to plot}
}
\description{
get/set chart object
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.name}}, \code{\link{chart.opt}},
  \code{\link{chart.plot}}, \code{\link{is.chart}}
}
