% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.opt}
\alias{chart.opt}
\alias{chart.opt<-}
\title{get/set the options for a chart}
\usage{
chart.opt(obj, name)

chart.opt(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of chart to plot}

\item{opt}{list of options for this chart}
}
\description{
get/set the options for a chart
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.name}}, \code{\link{chart.obj}},
  \code{\link{chart.plot}}, \code{\link{is.chart}}
}
