% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_class.R
\docType{methods}
\name{chart.plot,chart-method}
\alias{chart.plot,chart-method}
\title{plot chart}
\usage{
\S4method{chart.plot}{chart}(obj)
}
\arguments{
\item{obj}{a chart object}
}
\description{
plots a chart object
}
