% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.plot}
\alias{chart.plot}
\title{plot a chart for a dataset object}
\usage{
chart.plot(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a dataset object}
}
\description{
plot a chart for a dataset object
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.name}}, \code{\link{chart.obj}},
  \code{\link{chart.opt}}, \code{\link{is.chart}}
}
