% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_class.R, R/chart_node_class.R
\name{chart}
\alias{chart}
\alias{chart,chart_node-method}
\alias{chart<-}
\alias{chart<-,chart_node,struct_class-method}
\title{Constructor for struct chart objects}
\usage{
chart(object)

chart(object)

\S4method{chart}{chart_node}(object)

chart(object) <- value

\S4method{chart}{chart_node,struct_class}(object) <- value
}
\arguments{
\item{...}{named slots and their values that get passed to struct_class}
}
\value{
a chart object

a struct_class object

the chart object in the node

the modified node object
}
\description{
A base class in the \pkg{struct} package. Should not be called directly.
}
\details{
The \code{chart} class provides a template for figures, charts and plots
associated with other objects. For example, a DatasetExperiment object could have a
histogram plotted for a specified column.

Charts can have parameters but not outputs (other than the figure itself), as
chart objects are not intended to be used for calculations. The
\code{chart_plot} method can be used to display a chart for an object, and
\code{chart_names} can be used to list all chart objects associated with an
object.

Classes that inherit the \code{stato} class have STATO integration enabled, 
allowing stato_id to be set and formal names
and descriptions pulled from the STATO ontology database.
}
\examples{
C = example_chart()
C = example_chart()
NODE = chart_node(chart = C)
chart(NODE)
C = example_chart()
NODE = chart_node()
chart(NODE) = C
}
