% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\docType{class}
\name{example_chart-class}
\alias{example_chart-class}
\alias{example_chart}
\alias{chart.plot,example_chart,test_model-method}
\title{example chart object}
\usage{
\S4method{chart.plot}{example_chart,test_model}(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a test_model object}
}
\description{
an example of a chart object for documentation purposes
}
\examples{
C = example_chart()
chart.plot(C,test_model())
}
