% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart_build}
\alias{chart_build}
\title{Build a chart}
\usage{
chart_build(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a DatasetExperiment object}
}
\value{
the chart object with built data
}
\description{
Build a chart
}
