% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\docType{class}
\name{example_chart-class}
\alias{example_chart-class}
\alias{example_chart}
\alias{chart_plot,example_chart,example_model-method}
\title{example chart object}
\usage{
\S4method{chart_plot}{example_chart,example_model}(obj, dobj)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a example_model object}
}
\value{
a chart object
}
\description{
an example of a chart object for documentation purposes
}
\examples{
C = example_chart()
chart_plot(C,example_model())
}
