% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/struct_class.R
\name{chart_names}
\alias{chart_names}
\alias{chart_names,struct_class-method}
\title{chart names}
\usage{
chart_names(obj, ret = "char")

\S4method{chart_names}{struct_class}(obj, ret = "char")
}
\arguments{
\item{obj}{An object derived from the struct_class object}

\item{ret}{A string indicating whether a list of objects ('obj') or a list of chart 
names ('char') is returned. 'char' is default.}
}
\value{
list of chart names, or a list of chart objects
}
\description{
Returns a list of valid charts for a struct object
}
\details{
The chart_names method searches chart objects that specify the input object
type as an input.
}
\section{Methods (by class)}{
\itemize{
\item \code{struct_class}: 
}}

\examples{
M = example_model()
chart_names(M) # 'example_chart'
chart_names(M,'char') # as above
chart_names(M,'obj') # returns a list of chart objects
}
