% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_node_class.R
\name{chart_node}
\alias{chart_node}
\title{chart_node class}
\usage{
chart_node(name = character(0), description = character(0), chart = NULL, ...)
}
\arguments{
\item{name}{the name of the node}

\item{description}{a description of the node}

\item{chart}{a struct chart object}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A class for representing chart nodes in a directed acyclic graph (DAG) of struct models.
Each node contains a struct chart object that will be executed using chart_plot.
Chart nodes can have multiple input objects (e.g., a model and a dataset).
}
\section{\code{chart} slot}{

The "chart" slot holds a struct chart object that will be executed when the node is processed using chart_plot.
}

\section{\code{input_objects} slot}{

The "input_objects" slot holds a list of input objects that will be passed to chart_plot.
The order of objects in this list corresponds to the order of parameters in the chart_plot method.
}

\examples{
C = example_chart()
NODE = chart_node(name = 'Example Chart Node', description = 'A simple example chart node', chart = C)
}
