% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/chart_class.R
\name{chart_plot}
\alias{chart_plot}
\alias{chart_plot,chart,ANY-method}
\title{chart_plot}
\usage{
chart_plot(obj, dobj, ...)

\S4method{chart_plot}{chart,ANY}(obj, dobj)
}
\arguments{
\item{obj}{A chart object}

\item{dobj}{An object derived from struct_class}

\item{...}{optional inputs}
}
\value{
a plot object
}
\description{
Plots a chart object
}
\details{
The optional optional inputs depend on the input object/chart, but might 
include an additional dataset object or a second model object, for example.
}
\section{Methods (by class)}{
\itemize{
\item \code{obj = chart,dobj = ANY}: 
}}

\examples{
C = example_chart()
chart_plot(C,iris_DatasetExperiment())

}
