% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartable_class.R
\docType{class}
\name{chartable_class-class}
\alias{chartable_class-class}
\alias{chartable_class}
\title{chartable_class}
\description{
A base class in the \pkg{struct} package. Provides several fundamental methods for charting and should not be called directly.
}
