% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{data}
\alias{data}
\alias{data<-}
\title{get/set data for a dataset object}
\usage{
data(obj)

data(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of raw data}
}
\description{
get/set data for a dataset object
}
\seealso{
Other dataset functions: \code{\link{sample.meta}},
  \code{\link{variable.meta}}
}
