% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_node_class.R
\name{data_node}
\alias{data_node}
\title{data_node class}
\usage{
data_node(name = character(0), description = character(0), data = NULL, ...)
}
\arguments{
\item{name}{the name of the data node}

\item{description}{a description of the data node}

\item{data}{a DatasetExperiment object}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A class for representing data nodes in a directed acyclic graph (DAG) of struct models.
Each data node contains a DatasetExperiment object that can be used as input to other nodes.
}
\section{\code{data} slot}{

The "data" slot holds a DatasetExperiment object that will be used as input
to other nodes in the DAG.
}

\examples{
D = iris_DatasetExperiment()
DNODE = data_node(name = 'Example Data', description = 'A simple example data node', data = D)
}
