% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_class.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\alias{dataset.data,dataset-method}
\alias{dataset.data<-,dataset-method}
\alias{dataset.sample_meta,dataset-method}
\alias{dataset.sample_meta<-,dataset-method}
\alias{dataset.variable_meta,dataset-method}
\alias{dataset.variable_meta<-,dataset-method}
\title{dataset class}
\usage{
\S4method{dataset.data}{dataset}(obj)

\S4method{dataset.data}{dataset}(obj) <- value

\S4method{dataset.sample_meta}{dataset}(obj)

\S4method{dataset.sample_meta}{dataset}(obj) <- value

\S4method{dataset.variable_meta}{dataset}(obj)

\S4method{dataset.variable_meta}{dataset}(obj) <- value
}
\arguments{
\item{obj}{a dataset object}
}
\description{
an object for holding raw data and associated meta data
}
\section{Slots}{

\describe{
\item{\code{name}}{(character) name of the dataset}

\item{\code{description}}{(character) brief description of the dataset}

\item{\code{data}}{(data.frame) matrix of data, samples in rows, variables/features in columns.}

\item{\code{sample.meta}}{(data.frame) data frame of sample meta data e.g. group membership}

\item{\code{variable.meta}}{(data.frame) data frame of variable meta data}
}}

