% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{description}
\alias{description}
\alias{description<-}
\title{Object Description}
\usage{
description(obj)

description(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid string}
}
\value{
\describe{
\item{\code{description(obj)}}{returns the description of an object as a
string}
\item{\code{description(obj)<-}}{sets the description of an object and
returns the modified object}
}
}
\description{
Get/set the description of an object.
}
\examples{
M = example_model()
description(M)
description(M) = 'this is an example'
}
