% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{.initialize_struct_class}
\alias{.initialize_struct_class}
\title{Initialise a struct object}
\usage{
.initialize_struct_class(.Object, ...)
}
\arguments{
\item{.Object}{a prototype struct object}

\item{...}{named slots and their values.}
}
\value{
a struct object with populated slots
}
\description{
This function handles the conversion of params from name to params_name when 
initialising an object. It is used in the constructor for new struct objects, 
and not intended to be used in isolation.
}
\examples{
\dontrun{
# a class contructor can be created as follows
new_struct_class = function(...) {
    # new object
    out = .new_struct_class()
    # initialise
    out = .initialize_struct_class(out,...)
    return(out)
    }
}
}
