% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_class.R
\docType{class}
\name{entity-class}
\alias{entity-class}
\alias{entity}
\alias{value,entity-method}
\title{entity_class}
\usage{
\S4method{value}{entity}(MET)
}
\description{
A base class in the \pkg{struct} package. should not be called directly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{value}: get the calculated value for an entity
}}

