% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_class.R
\name{entity_constraint}
\alias{entity_constraint}
\title{Entity constraints}
\usage{
entity_constraint(ex, msg)
}
\arguments{
\item{...}{any number of constraints e.g. value<6}
}
\value{
A list of quosures that will be evaluated to check validity of the entity
}
\description{
This helper function creates a list of quosures that can be used to apply
constraints to an entity object.
}
\examples{
# ensure the value of an entity is exactly 6
constraint(value==6)

# check the length
constraint(length(value)==1)


}
