% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_stato_class.R
\docType{class}
\name{entity_stato-class}
\alias{entity_stato-class}
\alias{entity_stato}
\title{entity_stato class}
\description{
A base class in the \pkg{struct} package. Should not be called directly.
}
\details{
Extends the \code{entity} class to include stato functionality.
}
\examples{
E = entity_stato()
}
\seealso{
Refer to \code{\link{entity}} and \code{\link{stato}} for further
info.
}
