% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_class.R
\docType{class}
\name{enum-class}
\alias{enum-class}
\alias{enum}
\alias{value<-,enum-method}
\title{Enum objects}
\usage{
\S4method{value}{enum}(obj) <- value
}
\arguments{
\item{obj}{an enum object}

\item{value}{value of the enum}
}
\value{
an enum object
}
\description{
A base class in the \pkg{struct} package. Not normally called directly.
}
\details{
An enum object is a special type of entity object that ensures the value must
be one from a list of allowed values.

Enum objects are usually defined in the prototype of another object, but
can be extracted using \code{param.obj} and \code{output.obj}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{value<-}: set the value for an enum
}}

\examples{
# Create a new enum object
E = enum(
    name = 'example',
    description = 'this is an example',
    type = 'character',
    value = 'hello',
    list = c('hello','world')
)

# Get/set the value of the entity object
value(E)
value(E) = 'world'

}
