% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_stato_class.R
\docType{class}
\name{enum_stato-class}
\alias{enum_stato-class}
\alias{enum_stato}
\title{enum_stato class}
\description{
A base class in the \pkg{struct} package. Should not be called directly.
}
\details{
Extends the \code{enum} class to include stato functionality.
}
\examples{
E = enum_stato()
}
\seealso{
Refer to \code{\link{enum}} and \code{\link{stato}} for further
info.
}
