% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{example_dag}
\alias{example_dag}
\title{Example DAG}
\usage{
example_dag()
}
\value{
model_dag object
}
\description{
An example model_dag for testing the DAG chart functionality.
Creates a simple DAG with two connected nodes.
}
\examples{
DAG = example_dag()
C = model_dag_chart()
chart_plot(C, DAG)
}
