% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_class.R
\docType{class}
\name{example_iterator-class}
\alias{example_iterator-class}
\alias{example_iterator}
\alias{run,example_iterator,DatasetExperiment,metric-method}
\title{Example iterator}
\usage{
\S4method{run}{example_iterator,DatasetExperiment,metric}(I, D, MET)
}
\arguments{
\item{I}{example_iterator object}

\item{D}{dataset object}

\item{MET}{metric object}
}
\value{
test iterator object

dataset object
}
\description{
An example iterator for testing

runs the example iterator, which just returns a value of 3.142
}
\examples{
I = example_iterator()

I = example_iterator()
D = iris_DatasetExperiment()
MET = metric()
I = run(I,D,MET)

}
