% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{example_model}
\alias{example_model}
\alias{model_train,example_model,DatasetExperiment-method}
\alias{model_predict,example_model,DatasetExperiment-method}
\title{Example model}
\usage{
example_model(value_0 = 0, value_1 = 10, value_2 = 20, ...)

\S4method{model_train}{example_model,DatasetExperiment}(M, D)

\S4method{model_predict}{example_model,DatasetExperiment}(M, D)
}
\arguments{
\item{value_0}{a numeric value}

\item{value_1}{a numeric value}

\item{value_2}{a numeric value}

\item{...}{named slots and their values.}

\item{M}{A struct model object}

\item{D}{A DatasetExperiment object}
}
\value{
modified example_model object

dataset object

dataset object
}
\description{
An example model for testing. Training this model adds value_1 to a data set, and
prediction using this model adds value_2.

trains the example model, which adds value_1 to the raw data of a dataset

predicts using the example model, which adds value_2 to the raw data of a
dataset
}
\examples{
M = example_model()
M = example_model(value_1 = 10, value_2 = 20)
D = iris_DatasetExperiment()
M = example_model(value_1 = 10, value_2 = 20)
M = model_train(M,D)
D = iris_DatasetExperiment()
M = example_model(value_1 = 10, value_2 = 20)
M = model_predict(M,D)
}
