% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{export.xlsx}
\alias{export.xlsx}
\title{write a dataset object to file}
\usage{
export.xlsx(object, ...)
}
\arguments{
\item{object}{a dataset object}

\item{...}{additional input arguments (see description)}
}
\description{
write a dataset object to file
}
