% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DatasetExperiment_class.R
\name{export_xlsx}
\alias{export_xlsx}
\alias{export_xlsx,DatasetExperiment-method}
\title{write a dataset object to file}
\usage{
export_xlsx(object, outfile, transpose = TRUE)

\S4method{export_xlsx}{DatasetExperiment}(object, outfile, transpose = TRUE)
}
\arguments{
\item{object}{a dataset object}

\item{outfile}{the filename (including path) to write the data to}

\item{transpose}{TRUE (default) or FALSE to transpose the output data}
}
\value{
an excel file with sheets for data and meta data
}
\description{
Exports a dataset object to an excel file with sheets for data, sample_meta and variable_meta
}
\examples{
\dontrun{
D = iris_DatasetExperiment() # example dataset
export_xlsx(D,'iris_DatasetExperiment.xlsx')
}
}
