% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_description}
\alias{get_description}
\title{Get struct object help description}
\usage{
get_description(id)
}
\arguments{
\item{id}{(character) the name of a struct object to generate documentation for}
}
\value{
a character string of roxygen formatted documentation for the object
}
\description{
This
function is to help developers including struct objects in their own R
packages, and isnt intended for general use.
Use with roxygen 2 `@eval` tags this function generates a detailed
description of a struct object generated by extracting names,
descriptions etc from slots in a suitable format.
}
\examples{
get_description('example_model')

}
