% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/chart_class.R
\name{get_preset}
\alias{get_preset}
\alias{get_preset,chart,character,character-method}
\title{Get preset from layer entity or chart object}
\usage{
get_preset(obj, preset_name, slot_name)

\S4method{get_preset}{chart,character,character}(obj, preset_name, slot_name)
}
\arguments{
\item{obj}{A chart object}

\item{preset_name}{Character string naming the preset to retrieve}

\item{slot_name}{Character string naming the slot containing a layer_entity}
}
\value{
the preset configuration

The preset configuration
}
\description{
Get preset from layer entity or chart object

Get preset from chart object
}
\examples{
# Get a preset from a chart object
C = new_scatter_chart()
get_preset(C, 'default', 'points')
}
