% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{iris_dataset}
\alias{iris_dataset}
\title{fishers iris data}
\usage{
iris_dataset()
}
\value{
dataset object
}
\description{
fishers iris data as a struct dataset object
}
\examples{
D = iris_dataset()
summary(D)
}
