% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{iris_DatasetExperiment}
\alias{iris_DatasetExperiment}
\title{Fisher's Iris data}
\usage{
iris_DatasetExperiment()
}
\value{
DatasetExperiment object
}
\description{
Fisher's Iris data as a DatasetExperiment object
}
\examples{
D = iris_DatasetExperiment()
}
