% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.chart}
\alias{is.chart}
\title{verify chart name}
\usage{
is.chart(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of the chart as a string}
}
\value{
TRUE if chart name is valid, FALSE if not
}
\description{
verify that the name of chart is valid for an object
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.name}}, \code{\link{chart.obj}},
  \code{\link{chart.opt}}, \code{\link{chart.plot}}
}
