% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.output}
\alias{is.output}
\title{verify output name}
\usage{
is.output(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of the output as a string}
}
\value{
TRUE if output name is valid, FALSE if not
}
\description{
verify that the name of an output is valid for an object
}
\examples{
\dontrun{

is.input(M,'value_1')
}
}
