% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.param}
\alias{is.param}
\title{Verify parameter}
\usage{
is.param(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of parameter as a string}
}
\value{
TRUE if parameter name is valid, FALSE if not
}
\description{
Verify that the name of a paramater is valid for an object
}
\examples{
M = test_model()
is.param(M,'value_1') # TRUE
is.param(M,'alpha')   # FALSE

}
