% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{is_output}
\alias{is_output}
\alias{is_output,struct_class-method}
\title{Verify output}
\usage{
is_output(obj, name)

\S4method{is_output}{struct_class}(obj, name)
}
\arguments{
\item{obj}{A model or iterator object derived from the *struct* class}

\item{name}{Name of output}
}
\value{
TRUE if output name is valid, FALSE if not
}
\description{
Verify that the name of a output is valid for an object
}
\section{Methods (by class)}{
\itemize{
\item \code{struct_class}: 
}}

\examples{
M = example_model()
is_output(M,'result_1') # TRUE
is_output(M,'result_0')   # FALSE
}
