% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_class.R
\docType{class}
\name{iterator-class}
\alias{iterator-class}
\alias{iterator}
\alias{run,iterator,dataset,metric-method}
\alias{evaluate,iterator,metric-method}
\alias{models,iterator-method}
\alias{models<-,iterator,model_OR_iterator-method}
\alias{models<-,iterator,model_OR_iterator-method}
\alias{result,iterator-method}
\alias{result.name,iterator-method}
\title{iterator class}
\usage{
\S4method{run}{iterator,dataset,metric}(I, D, MET = NULL)

\S4method{evaluate}{iterator,metric}(I, MET)

\S4method{models}{iterator}(ML)

\S4method{models}{iterator,model_OR_iterator}(ML) <- value

\S4method{models}{iterator,model_OR_iterator}(ML) <- value

\S4method{result}{iterator}(M)

\S4method{result.name}{iterator}(M)
}
\arguments{
\item{I}{an iterator object}

\item{D}{a dataset object}

\item{MET}{a metric object}
}
\description{
A class for iterative approaches that ivolve the training/prediction of a model multiple times.
not intended to be called directly, this class should be inherited to provide functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run}: run a model/model.list mutliple times for the input data

\item \code{evaluate}: evaluate the performance of a model/model.list  using the input metric

\item \code{models}: get the model/model.list for an iterator object

\item \code{models<-}: set the model/model.list to be run for multiple iterations

\item \code{models<-}: set the model/model.list to be run for multiple iterations

\item \code{result}: get prediction output from iterator

\item \code{result.name}: get prediction output name for iterator
}}

