% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_entity_class.R
\name{layer_entity}
\alias{layer_entity}
\alias{register_preset<-}
\alias{available_presets}
\alias{value<-,layer_entity-method}
\alias{$<-,layer_entity-method}
\alias{show,layer_entity-method}
\alias{as_layer}
\alias{value,layer_entity-method}
\alias{get_preset,layer_entity,character,missing-method}
\title{Layer Entity objects}
\usage{
layer_entity(
  layer = "geom_point",
  presets = list(),
  preset_default = "default",
  value = preset("default"),
  ...
)

register_preset(obj, preset_name, force = TRUE) <- value

available_presets(obj, name)

\S4method{value}{layer_entity}(obj) <- value

\S4method{$}{layer_entity}(x, name) <- value

\S4method{show}{layer_entity}(object)

as_layer(obj, ...)

\S4method{value}{layer_entity}(obj)

\S4method{get_preset}{layer_entity,character,missing}(obj, preset_name, slot_name)
}
\arguments{
\item{layer}{Character string naming the ggplot2 layer function (e.g., 'geom_point')}

\item{presets}{List of preset configurations}

\item{preset_default}{Character string naming the default preset}

\item{value}{The value of the layer entity}

\item{...}{Additional parameters passed to entity constructor}
}
\value{
A layer_entity object
}
\description{
A class in the \pkg{struct} package for managing ggplot2 layers with preset configurations.
Layer entities extend the base entity class to provide functionality for creating
and managing ggplot2 layers with predefined presets.
}
\examples{
# Create a layer entity
le = layer_entity(
    layer = 'geom_point',
    presets = list(
        default = list(color = 'red'),
        blue = list(color = 'blue')
    ),
    preset_default = 'default',
    value = preset('default')
)
}
