% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{list_struct}
\alias{list_struct}
\alias{as_list_struct}
\alias{un_list_struct}
\alias{is_list_struct}
\alias{tibble_struct}
\alias{as_tibble_struct}
\alias{un_tibble_struct}
\alias{is_tibble_struct}
\alias{scalar}
\alias{un_scalar}
\alias{is_scalar}
\title{struct classes}
\usage{
list_struct(...)

as_list_struct(x)

un_list_struct(x)

is_list_struct(x)

tibble_struct(...)

as_tibble_struct(x)

un_tibble_struct(x)

is_tibble_struct(x)

scalar(x)

un_scalar(x)

is_scalar(x)
}
\arguments{
\item{...}{elements of the object to build}

\item{x}{object}
}
\description{
Functions to convert to and from struct classes or to test objects.
}
