% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_class.R
\docType{class}
\name{method-class}
\alias{method-class}
\alias{method}
\alias{method.apply,method,dataset-method}
\alias{predicted,method-method}
\alias{predicted.name,method-method}
\alias{predicted.name<-,method,character-method}
\title{method class}
\usage{
\S4method{method.apply}{method,dataset}(M, D)

\S4method{predicted}{method}(M)

\S4method{predicted.name}{method}(M)

\S4method{predicted.name}{method,character}(M) <- value
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}

\item{value}{value}
}
\value{
a method object
}
\description{
A class for methods that are applied to datasets in a single step e.g.
normalisation. Not intended to be called directly, this class should be
inherited to provide functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{method.apply}: train the model using input data

\item \code{predicted}: get prediction output from model

\item \code{predicted.name}: get prediction output name for model

\item \code{predicted.name<-}: set prediction output from method
}}

\examples{
M = method() # create a method object

D = iris_dataset() # example dataset
M = method.apply(M,D) # apply method to data

\dontrun{

p = predicted(M) # get the predicted output from the model
}

\dontrun{

M = method()
predicted.name(M)
}

M = method()
predicted.name(M) = 'example'

}
