% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{method.apply}
\alias{method.apply}
\title{Apply a method}
\usage{
method.apply(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified method object
}
\description{
Applies a method to the input dataset
}
\examples{
M = example_method()
M = method.apply(M,iris_dataset())

}
