% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_stato_class.R
\docType{class}
\name{method.stato-class}
\alias{method.stato-class}
\alias{method.stato}
\title{stato_method_class}
\description{
A base class in the \pkg{struct} package. should not be called directly.
}
\examples{
M = method.stato()
}
