% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{method.steps}
\alias{method.steps}
\alias{method.steps<-}
\title{Get/set methods of a method.seq}
\usage{
method.steps(ML)

method.steps(ML) <- value
}
\arguments{
\item{ML}{a method.seq object}

\item{value}{a list containing only method objects}
}
\value{
\describe{
\item{\code{method.steps(ML)}}{returns a list of methods}
\item{\code{method.steps(ML)<-}}{sets the list of methods for a sequence}
}
}
\description{
Returns the list of method in a method.seq object
}
\examples{
ML = example_model() + example_model()
method.steps(ML)
method.steps(ML) = list(example_model(),example_model())

}
