% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_class.R
\docType{class}
\name{metric-class}
\alias{metric-class}
\alias{metric}
\alias{calculate,metric-method}
\alias{value,metric-method}
\alias{value<-,metric-method}
\title{metric class}
\usage{
\S4method{calculate}{metric}(obj, Y, Yhat)

\S4method{value}{metric}(obj)

\S4method{value}{metric}(obj) <- value
}
\arguments{
\item{obj}{a metric object}

\item{Y}{the true class labels}

\item{Yhat}{the predicted class labels}

\item{value}{value}
}
\value{
a metric object
}
\description{
A class for metrics to assess performance of e.g. models, iterators.
Not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{calculate}: calculate a metric

\item \code{value}: get the caluclated value for a metric

\item \code{value<-}: set the caluclated value for a metric
}}

\examples{
MET = metric()
M = metric()
calculate(M,Y,Yhat)
MET = metric()
value(MET)
MET = metric()
value(MET)=10
}
