% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calculate}
\alias{calculate}
\alias{value}
\alias{value<-}
\title{Calculate metric}
\usage{
calculate(obj, ...)

value(obj)

value(obj) <- value
}
\arguments{
\item{obj}{a metric object}

\item{...}{additional inputs depending on object}

\item{value}{value}
}
\value{
value the calculated value of a metric
}
\description{
Calculate metric

get the value for a metric

set the value for a metric
}
\examples{
MET = metric()
calculate(MET)

}
