% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stato_class.R
\docType{class}
\name{model.stato-class}
\alias{model.stato-class}
\alias{model.stato}
\title{stato_model_class}
\description{
A base class in the \pkg{struct} package. should not be called directly.
}
