% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_apply}
\alias{model_apply}
\title{Apply a model}
\usage{
model_apply(M, D)
}
\arguments{
\item{M}{a method object}

\item{D}{another object used by the first}
}
\value{
Returns a modified method object
}
\description{
Applies a method to the input dataset
}
\examples{
M = example_model()
M = model_apply(M,iris_DatasetExperiment())
}
