% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_node_class.R
\name{model_node}
\alias{model_node}
\title{model_node class}
\usage{
model_node(name = character(0), description = character(0), model = NULL, ...)
}
\arguments{
\item{name}{the name of the node}

\item{description}{a description of the node}

\item{model}{a struct model object}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A class for representing nodes in a directed acyclic graph (DAG) of struct models.
Each node contains a struct model object that will be executed using model_apply.
}
\section{\code{model} slot}{

The "model" slot holds a struct model object (e.g., PCA, PLS, etc.) that will be
executed when the node is processed using model_apply.
}

\examples{
M = example_model()
NODE = model_node(name = 'Example Node', description = 'A simple example node', model = M)
}
