% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_reverse}
\alias{model_reverse}
\title{Reverse preprocessing}
\usage{
model_reverse(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified dataset object
}
\description{
Reverse the effect of a preprocessing step on a dataset_
}
\examples{
M = example_model()
D = model_reverse(M,iris_dataset())

}
