% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{models}
\alias{models}
\alias{models<-}
\title{Get/set models of a model.seq}
\usage{
models(ML)

models(ML) <- value
}
\arguments{
\item{ML}{a model.seq object}

\item{value}{a list containing only model objects}
}
\description{
Returns the list of models in a model.seq object
}
