% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{name}
\alias{name}
\alias{name<-}
\title{Object Name}
\usage{
name(obj)

name(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid string}
}
\value{
\describe{
\item{\code{name(obj)}}{returns the name of an object as a string}
\item{\code{name(obj)<-}}{sets the name of an object and returns the modified
object}
}
}
\description{
Get/set the name of an object. This is the long name for an object, not the
name of the class object.
}
\examples{
M = example_model()
name(M)
name(M) = 'example object'
}
