% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{new_struct}
\alias{new_struct}
\title{Generate a \pkg{struct} object from a Class}
\usage{
new_struct(class, ...)
}
\arguments{
\item{class}{The class of struct object to create}

\item{...}{named slots and values to assign}
}
\value{
An object derived from struct_class
}
\description{
This function creates a newly allocated object from the class identified by 
the first argument. It works almost identically to \code{new} but is specific 
to objects from the \pkg{struct} package and ensures that \code{entity} slots have 
their values assigned correctly. This function is usually called by class 
constructors and not used directly.
}
\examples{
S = new_struct('struct_class')
}
