% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\docType{methods}
\name{description,struct_class-method}
\alias{description,struct_class-method}
\alias{description<-,struct_class-method}
\title{description}
\usage{
\S4method{description}{struct_class}(obj)

\S4method{description}{struct_class}(obj) <- value
}
\arguments{
\item{obj}{a struct object}

\item{value}{value}
}
\description{
get the description of an object
}
