% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{name,struct_class-method}
\alias{name,struct_class-method}
\alias{name<-,struct_class-method}
\title{name}
\usage{
\S4method{name}{struct_class}(obj)

\S4method{name}{struct_class}(obj) <- value
}
\arguments{
\item{obj}{a struct object}

\item{value}{value}
}
\value{
(long) name of struct object

modified struct object
}
\description{
get the name of an object
}
\examples{
S = struct_class()
name(S)
S = struct_class()
name(S) = 'example'
}
