% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiser_class.R
\docType{class}
\name{optimiser-class}
\alias{optimiser-class}
\alias{optimiser}
\title{optimiser class}
\description{
A special class of iterator for selecting optimal parameter values
not intended to be called directly, this class should be inherited to provide functionality for method-specific classes.
}
