% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiser_class.R
\name{optimiser}
\alias{optimiser}
\title{optimiser class}
\usage{
optimiser(...)
}
\arguments{
\item{...}{named slots and their values.}
}
\value{
an optimiser object
}
\description{
A special class of iterator for selecting optimal parameter values
not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\examples{
OPT = optimiser()
}
