% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.list}
\alias{output.list}
\alias{output.list<-}
\title{output list}
\usage{
output.list(obj)

output.list(obj) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{value}{a named list of outputs for an object}
}
\value{
list of output names
}
\description{
get a named list of outputs and thier current value for an object
}
\seealso{
Other output functions: \code{\link{is.output}},
  \code{\link{output.ids}}, \code{\link{output.obj}},
  \code{\link{output.value}}
}
\concept{output functions}
