% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.name}
\alias{output.name}
\title{output names}
\usage{
output.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{id of output}
}
\value{
list of parameter names
}
\description{
return a list of valid parameter names for an object
}
\examples{
\dontrun{

output.name(M,'value_1')
}
}
