% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.obj}
\alias{output.obj}
\alias{output.obj<-}
\title{Output objects}
\usage{
output.obj(obj, name)

output.obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of output as a string}

\item{value}{a valid value for the output being set}
}
\value{
\describe{
\item{\code{output.obj(M,name)}}{returns the named output as an object}
\item{\code{output.obj(M,name)<-}}{sets the named output of an object}
}
}
\description{
Gets or sets the object of an output e.g. to an entity() object.
}
\examples{
# get the output as an object
M = example_model()
obj = output.obj(M, 'result_1')

# set a output as an object
output.obj(M, 'result_1') = entity(value = 15,type='numeric')

}
