% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.value}
\alias{output.value}
\alias{output.value<-}
\title{output values}
\usage{
output.value(obj, name)

output.value(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of an output}

\item{value}{value}
}
\value{
value of output
}
\description{
get/set the values for an output.
}
\seealso{
Other output functions: \code{\link{is.output}},
  \code{\link{output.ids}}, \code{\link{output.list}},
  \code{\link{output.obj}}
}
