% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{output_ids}
\alias{output_ids}
\alias{output_ids,struct_class-method}
\title{Output identifiers}
\usage{
output_ids(obj)

\S4method{output_ids}{struct_class}(obj)
}
\arguments{
\item{obj}{A model or iterator object derived from the *struct* class}
}
\value{
list of output ids
}
\description{
return a list of valid output ids for an object
}
\examples{
M = example_model()
output_ids(M)
}
