% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{output_list}
\alias{output_list}
\alias{output_list<-}
\alias{output_list,struct_class-method}
\alias{output_list<-,struct_class,list-method}
\title{output list}
\usage{
output_list(obj)

output_list(obj) <- value

\S4method{output_list}{struct_class}(obj)

\S4method{output_list}{struct_class,list}(obj) <- value
}
\arguments{
\item{obj}{An object derived from struct_class}

\item{value}{A named list of outputs and corresponding values}
}
\value{
A named list of outputs and corresponding values

struct object
}
\description{
get/set a named list of outputs and their current value for an object
}
\section{Methods (by class)}{
\itemize{
\item \code{struct_class}: 

\item \code{obj = struct_class,value = list}: 
}}

\examples{
M = example_model()
L = output_list(M)
M = example_model()
output_list(M) = list('result_1' = DatasetExperiment(),'result_2' = DatasetExperiment())
}
