% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{output_name}
\alias{output_name}
\alias{output_name,struct_class,character-method}
\title{output name}
\usage{
output_name(obj, name)

\S4method{output_name}{struct_class,character}(obj, name)
}
\arguments{
\item{obj}{A model or iterator object derived from the *struct* class}

\item{name}{Name of output}
}
\value{
name of output
}
\description{
return a the name for a output, if available
}
\section{Methods (by class)}{
\itemize{
\item \code{obj = struct_class,name = character}: 
}}

\examples{
M = example_model()
output_name(M,'result_1')
}
