% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{output_obj}
\alias{output_obj}
\alias{output_obj<-}
\alias{output_obj,struct_class,character-method}
\alias{output_obj<-,struct_class,character-method}
\title{Output objects}
\usage{
output_obj(obj, name)

output_obj(obj, name) <- value

\S4method{output_obj}{struct_class,character}(obj, name)

\S4method{output_obj}{struct_class,character}(obj, name) <- value
}
\arguments{
\item{obj}{A model or iterator object derived from the *struct* class}

\item{name}{Name of output}

\item{value}{A valid value for the output being set}
}
\value{
\describe{
\item{\code{output_obj(M,name)}}{returns the named output as an object}
\item{\code{output_obj(M,name)<-}}{sets the named output of an object}
}

the modified object
}
\description{
Gets or sets the object of an output e.g. to an entity() object.
}
\examples{
# get the output as an object
M = example_model()
obj = output_obj(M, 'result_1')

# set a output as an object
output_obj(M, 'result_1') = entity(value = 15,type = 'numeric',name = 'result_1')
}
