% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/output_class.R
\name{output_value}
\alias{output_value}
\alias{output_value<-}
\alias{output_value,struct_class,character-method}
\alias{output_value<-,struct_class,character-method}
\title{output values}
\usage{
output_value(obj, name)

output_value(obj, name) <- value

\S4method{output_value}{struct_class,character}(obj, name)

\S4method{output_value}{struct_class,character}(obj, name) <- value
}
\arguments{
\item{obj}{A model or iterator object derived from the *struct* class}

\item{name}{Name of output}

\item{value}{A valid value for the output being set}
}
\value{
Value of output

struct object
}
\description{
get/set the values for an output_
}
\section{Methods (by class)}{
\itemize{
\item \code{obj = struct_class,name = character}: 

\item \code{obj = struct_class,name = character}: 
}}

\examples{
M = example_model()
output_value(M,'result_1')
M = example_model()
output_value(M,'result_1') = DatasetExperiment()
}
