% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.ids}
\alias{param.ids}
\title{parameter identifiers}
\usage{
param.ids(obj)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
list of parameter ids
}
\description{
return a list of valid parameter ids for an object
}
\seealso{
Other parameter functions: \code{\link{is.param}},
  \code{\link{output.name}}, \code{\link{param.list}},
  \code{\link{param.name}}, \code{\link{param.obj}},
  \code{\link{param.value}}
}
