% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.name}
\alias{param.name}
\title{parameter name}
\usage{
param.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{id of parameter}
}
\value{
name of parameter
}
\description{
return a the name for a paramater, if available
}
\examples{
M = example_model()
param.name(M,'value_1')

}
