% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.value}
\alias{param.value}
\alias{param.value<-}
\title{parameter values}
\usage{
param.value(obj, name)

param.value(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{of a parameter}

\item{value}{value}
}
\value{
value of output
}
\description{
get/set the values for a parameter.
}
\seealso{
Other parameter functions: \code{\link{is.param}},
  \code{\link{output.name}}, \code{\link{param.ids}},
  \code{\link{param.list}}, \code{\link{param.name}},
  \code{\link{param.obj}}
}
