% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/parameter_class.R
\name{param_ids}
\alias{param_ids}
\alias{param_ids,struct_class-method}
\title{Parameter identifiers}
\usage{
param_ids(obj)

\S4method{param_ids}{struct_class}(obj)
}
\arguments{
\item{obj}{An object derived from struct_class}
}
\value{
list of parameter ids
}
\description{
return a list of valid parameter ids for an object
}
\examples{
M = example_model()
param_ids(M)
}
