% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/parameter_class.R
\name{param_list}
\alias{param_list}
\alias{param_list<-}
\alias{param_list,struct_class-method}
\alias{param_list<-,struct_class,list-method}
\title{Parameter list}
\usage{
param_list(obj)

param_list(obj) <- value

\S4method{param_list}{struct_class}(obj)

\S4method{param_list}{struct_class,list}(obj) <- value
}
\arguments{
\item{obj}{An object derived from struct_class}

\item{value}{A named list of parameters and corresponding values}
}
\value{
A named list of parameters names and corresponding values
}
\description{
get/set a named list of parameters and thier current value for an object
}
\examples{
M = example_model()
L = param_list(M)

M = example_model()
param_list(M) = list('value_1' = 15,'value_2' = 20)
}
