% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/parameter_class.R
\name{param_name}
\alias{param_name}
\alias{param_name,struct_class,character-method}
\title{Parameter name}
\usage{
param_name(obj, name)

\S4method{param_name}{struct_class,character}(obj, name)
}
\arguments{
\item{obj}{An object derived from struct_class}

\item{name}{Name of parameter}
}
\value{
name of parameter
}
\description{
Returns the name for a parameter, if available
}
\section{Methods (by class)}{
\itemize{
\item \code{obj = struct_class,name = character}: 
}}

\examples{
M = example_model()
param_name(M,'value_1')
}
