% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/parameter_class.R
\name{param_value}
\alias{param_value}
\alias{param_value<-}
\alias{param_value,struct_class,character-method}
\alias{param_value<-,struct_class,character-method}
\title{Parameter values}
\usage{
param_value(obj, name)

param_value(obj, name) <- value

\S4method{param_value}{struct_class,character}(obj, name)

\S4method{param_value}{struct_class,character}(obj, name) <- value
}
\arguments{
\item{obj}{A model or iterator object derived from structclass}

\item{name}{Name of parameter}

\item{value}{A valid value for the parameter being set}
}
\value{
Value of parameter
}
\description{
get/set the values for a parameter.
}
\examples{
M = example_model()
param_value(M,'value_1')

M = example_model()
param_value(M,'value_1') = 0.95

}
