% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_class.R
\docType{class}
\name{parameter_class-class}
\alias{parameter_class-class}
\alias{parameter_class}
\title{parameter_class}
\description{
A base class in the \pkg{struct} package. Provides several fundamental methods for getting/setting parameters etc and should not be called directly.
}
