% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model.predict}
\alias{model.predict}
\title{Apply a (trained) model to a dataset D to get a prediction}
\usage{
model.predict(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\description{
Apply a model using the input dataset. Requires that the model is trained first.
}
\seealso{
Other model functions: \code{\link{model.train}},
  \code{\link{predicted.name}}, \code{\link{predicted}}
}
\concept{model functions}
