% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predicted.name}
\alias{predicted.name}
\alias{predicted.name<-}
\title{get/set output name as prediction output for a model}
\usage{
predicted.name(M)

predicted.name(M) <- value
}
\arguments{
\item{M}{a model object}

\item{value}{name of an output for this model}
}
\description{
get/set the prediction output for a model. This determines which outputs from this model are supplied as inputs to the next model when used in a model.seq
}
\seealso{
Other model functions: \code{\link{model.predict}},
  \code{\link{model.train}}, \code{\link{predicted}}
}
\concept{model functions}
