% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predicted}
\alias{predicted}
\title{prediction output for a model}
\usage{
predicted(M)
}
\arguments{
\item{M}{a model object}
}
\description{
returns the prediction output for a model. This is supplied as input to the
next model when used in a model.seq
}
