% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predicted_name}
\alias{predicted_name}
\alias{predicted_name<-}
\title{Predicted output name}
\usage{
predicted_name(M)

predicted_name(M) <- value
}
\arguments{
\item{M}{a model object}

\item{value}{name of an output for this model}
}
\value{
\describe{
\item{\code{predicted_name}}{returns the name of the predicted output}
\item{\code{predicted_name<-}}{sets the name of the predicted output and
returns the modified object}
}
}
\description{
get/set the prediction output for a model_ This determines which outputs from
this model are supplied as inputs to the next model when used in a model_seq
}
\examples{
M = example_model()
predicted_name(M)
predicted_name(M) = 'result_2'
}
