% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_node_class.R
\name{prediction_node}
\alias{prediction_node}
\title{prediction_node class}
\usage{
prediction_node(name = character(0), description = character(0), ...)
}
\arguments{
\item{name}{the name of the node}

\item{description}{a description of the node}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A class for representing prediction nodes in a directed acyclic graph (DAG) of struct models.
Each prediction node receives a trained model via an edge and executes model_predict.
}
\examples{
NODE = prediction_node(name = 'Example Prediction', description = 'A simple example prediction node')
}
