% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_class.R
\docType{class}
\name{preprocess-class}
\alias{preprocess-class}
\alias{preprocess}
\alias{preprocess.reverse,preprocess,dataset-method}
\title{preprocessing class}
\usage{
\S4method{preprocess.reverse}{preprocess,dataset}(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}

\item{value}{value}
}
\value{
dataset object
}
\description{
A class used for preprocessing steps that require application to test sets.
not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{preprocess.reverse}: reverse the preprocessing step applied. Used in cases
such as mean centring, where you may want to 'uncentre' the results
}}

\examples{
M = preprocess()
D = dataset()
M = model()
D2 = preprocess.reverse(M,D)
}
