% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{preprocess.reverse}
\alias{preprocess.reverse}
\title{Reverse preprocessing}
\usage{
preprocess.reverse(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified dataset object
}
\description{
Reverse the effect of a preprocessing step on a dataset.
}
\examples{
M = example_model()
D = preprocess.reverse(M,iris_dataset())

}
