% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_class.R
\name{preprocess}
\alias{preprocess}
\alias{model_reverse,preprocess,dataset-method}
\title{preprocessing class}
\usage{
preprocess(...)

\S4method{model_reverse}{preprocess,dataset}(M, D)
}
\arguments{
\item{...}{named slots and their values.}

\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
dataset object
}
\description{
A class used for preprocessing steps that require application to test sets.
not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\examples{
M = preprocess()
D = dataset()
M = model()
D2 = model_reverse(M,D)
}
