% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_tree}
\alias{print_tree}
\title{print tree structure of object}
\usage{
print_tree(x)
}
\arguments{
\item{x}{object}
}
\description{
Not quite right yet. We want to be able to check easily the structure of an
object. Doesn't dive deep into lists unless they are named. Ultimately should
dive into the list_of ptypes though, and look prettier.
}
