% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{register_preset<-}
\alias{register_preset<-}
\title{Register a preset for a layer entity}
\usage{
register_preset(obj, preset_name, force = TRUE) <- value
}
\arguments{
\item{obj}{a layer_entity object}

\item{preset_name}{character string naming the preset}

\item{value}{list containing the preset configuration}

\item{force}{logical, whether to force overwrite existing preset}
}
\value{
the layer_entity object
}
\description{
Register a preset for a layer entity
}
