% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampler_class.R
\docType{class}
\name{resampler-class}
\alias{resampler-class}
\alias{resampler}
\title{resampler class}
\description{
A class for resampling methods such as cross-validation. not intended to be
called directly.
}
\examples{
R = resampler()
}
