% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampler_class.R
\name{resampler}
\alias{resampler}
\title{resampler class}
\usage{
resampler(...)
}
\arguments{
\item{...}{named slots and their values.}
}
\value{
a resampler object
}
\description{
A class for resampling methods such as cross-validation. not intended to be
called directly.
}
\examples{
R = resampler()
}
