% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{result.name}
\alias{result.name}
\alias{result.name<-}
\title{get/set output name as prediction output for a model}
\usage{
result.name(M)

result.name(M) <- value
}
\arguments{
\item{M}{an iterator object}

\item{value}{name of an output for iterator M}
}
\description{
get/set the prediction output for a model. This determines which outputs from this model are supplied as inputs to the next model when used in a model.seq
}
