% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{result_name}
\alias{result_name}
\alias{result_name<-}
\title{get/set output name as prediction output for a model}
\usage{
result_name(M)

result_name(I) <- value
}
\arguments{
\item{M}{an iterator object}

\item{I}{an iterator object}

\item{value}{name of an output for iterator M}
}
\value{
\describe{
\item{\code{result_name(M)}}{returns the name of the output for this iterator
(equivalent to \code{predicted} for model objects)}
\item{\code{result_name(I)<-}}{sets the default output for an iterator}
}
}
\description{
get/set the prediction output for a model_ This determines which outputs
from this model are supplied as inputs to the next model when used in
a model_seq
}
\examples{
I = example_iterator() # initialise iterator
result_name(I)
result_name(I) = 'result_1'
}
