% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sample.meta}
\alias{sample.meta}
\alias{sample.meta<-}
\title{get/set sample meta data for a dataset object}
\usage{
sample.meta(obj)

sample.meta(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of meta data}
}
\description{
get/set sample meta data for a dataset object
}
\seealso{
Other dataset functions: \code{\link{data}},
  \code{\link{variable.meta}}
}
