% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset.sample_meta}
\alias{dataset.sample_meta}
\alias{dataset.sample_meta<-}
\title{Sample meta data}
\usage{
dataset.sample_meta(obj)

dataset.sample_meta(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of meta data}
}
\value{
\describe{
\item{\code{dataset.sample_meta(obj)}}{returns a data.frame of meta data}
\item{\code{dataset.sample_meta(obj)<-}}{sets the meta data of a dataset
object}
}
}
\description{
get/set sample meta data for a dataset object
}
\examples{
# iris dataset
D = iris_dataset()
X = dataset.sample_meta(D)
dataset.sample_meta(D) = X
}
