% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\name{sbcms_dataset}
\alias{sbcms_dataset}
\title{data from sbcms}
\usage{
sbcms_dataset()
}
\value{
dataset object
}
\description{
sbcms data as a struct dataset object
}
\examples{
D = sbcms_dataset()
summary(D)
}
