% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{seq_in}
\alias{seq_in}
\alias{seq_in<-}
\title{Sequence input}
\usage{
seq_in(M)

seq_in(M) <- value
}
\arguments{
\item{M}{a model object}

\item{value}{name of an output for this model}
}
\value{
\describe{
\item{\code{seq_in}}{returns the name of the input parameter replaced
when used in a model sequence}
\item{\code{seq_in<-}}{sets the name of the input parameter replaced
when used in a model sequence}
}
}
\description{
get/set the input parameter replaced by the output of the previous model in
a model sequence. Default is "data" which passes the output as the data input
for methods such as \code{model_train} and \code{model_apply}.
}
\examples{
M = example_model()
seq_in(M)
seq_in(M) = 'value_1'
}
