% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{set_struct_obj}
\alias{set_struct_obj}
\title{define a new struct object}
\usage{
set_struct_obj(
  class_name,
  struct_obj,
  stato = TRUE,
  params = character(0),
  outputs = character(0),
  private = character(0),
  prototype = list()
)
}
\arguments{
\item{class_name}{the name of the new class to create}

\item{struct_obj}{the struct obj to inherit e.g. 'model', 'metric' etc}

\item{stato}{TRUE (default) or FALSE to inherit the stato class}

\item{params}{a named character vector of input parameters where each
element specifies the type of value that will be in the slot e.g. c(example = 'character')}

\item{outputs}{a named character vector of outputs where each
element specifies the type of value that will be in the slot e.g. c(example = 'character')}

\item{private}{a named character vector of private slots where each
element specifies the type of value that will be in the slot e.g. c(example = 'character').
These are intended for internal use by the object and generally not available to the user.}

\item{prototype}{a named list with initial values for slots.}
}
\value{
a new class definition. to create  a new object from this class use X = new_class_name()
}
\description{
a helper function to create new struct objects
}
