% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\docType{methods}
\name{show,struct_class-method}
\alias{show,struct_class-method}
\title{object summary}
\usage{
\S4method{show}{struct_class}(object)
}
\description{
print a brief summary of a struct object
}
