% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stato_class.R
\docType{class}
\name{stato-class}
\alias{stato-class}
\alias{stato}
\alias{stato.id,stato-method}
\alias{stato.name,stato-method}
\alias{stato.definition,stato-method}
\alias{stato.summary,stato-method}
\title{stato_class}
\usage{
\S4method{stato.id}{stato}(obj)

\S4method{stato.name}{stato}(obj)

\S4method{stato.definition}{stato}(obj)

\S4method{stato.summary}{stato}(obj)
}
\arguments{
\item{obj}{stato_class object}
}
\value{
stato id

stato name

stato definition

summary of the stato object
}
\description{
A base class in the \pkg{struct} package. Provides several fundamental
methods for the STATO ontology and should not be called directly.
}
\section{Methods (by generic)}{
\itemize{
\item \code{stato.id}: get the stato.id for an object

\item \code{stato.name}: get the STATO name for an object

\item \code{stato.definition}: get the STATO definition for an object

\item \code{stato.summary}: get the STATO definition for an object
}}

\examples{
S = stato()
\dontrun{
M = model.stato()
stato.id(M)
}
\dontrun{
M = model()
stato.name(M)
}
\dontrun{
M = model()
stato.definition(M)
}
\dontrun{
M = model()
stato.summary(M,'example')
}
}
