% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{stato_id}
\alias{stato_id}
\alias{stato_name}
\alias{stato_definition}
\alias{stato_summary}
\title{get the stato_id for an object}
\usage{
stato_id(obj)

stato_name(obj)

stato_definition(obj)

stato_summary(obj)
}
\arguments{
\item{obj}{stato_class object}
}
\value{
id the stato id

name the stato name

def the stato description
}
\description{
get the stato_id for an object

get the stato name for an object

get the stato name for an object

display a stato summary of the object
}
\examples{
M = example_model()
stato_id(M)
stato_name(M)
stato_definition(M)
stato_summary(M)

}
