% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct-package.R
\docType{package}
\name{struct-package}
\alias{struct}
\alias{struct-package}
\title{struct: Statistics in R Using Class-based Templates}
\description{
Defines and includes a set of class-based templates for developing and implementing data processing and analysis workflows, with a strong emphasis on statistics and machine learning. The templates can be used and where needed extended to 'wrap' tools and methods from other packages into a common standardised structure to allow for effective and fast integration. Model objects can be combined into sequences, and sequences nested in iterators using overloaded operators to simplify and improve readability of the code. Ontology lookup has been integrated and implemented to provide standardised definitions for methods, inputs and outputs wrapped using the class-based templates.
}
\author{
\strong{Maintainer}: Gavin Rhys Lloyd \email{g.r.lloyd@bham.ac.uk}

Authors:
\itemize{
  \item Ralf Johannes Maria Weber \email{r.j.weber@bham.ac.uk}
}

}
\keyword{internal}
