% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\docType{class}
\name{struct_class-class}
\alias{struct_class-class}
\alias{struct_class}
\title{struct_class}
\description{
A base class in the \pkg{struct} package. Provides several fundamental methods for getting/setting parameters etc and should not be called directly.
}
\examples{
S = struct_class()
}
