% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{struct_class}
\alias{struct_class}
\title{Constructor for struct_class objects}
\usage{
struct_class(
  name = character(0),
  description = character(0),
  type = character(0),
  citations = list(),
  ontology = character(0)
)
}
\arguments{
\item{name}{the name of the object}

\item{description}{a description of the object}

\item{type}{the type of the struct object}

\item{citations}{a list of citations for the object in "bibentry" format}

\item{ontology}{a list of ontology ids for the object}
}
\value{
a struct_class object
}
\description{
Creates a new \linkS4class{struct_class} object and populates the slots. Not intended
for direct use.
}
