% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_class.R
\name{dataset}
\alias{dataset}
\alias{$,dataset-method}
\alias{$<-,dataset-method}
\alias{summary,dataset-method}
\title{Dataset class}
\usage{
dataset(...)

\S4method{$}{dataset}(x, name)

\S4method{$}{dataset}(x, name) <- value

\S4method{summary}{dataset}(object)
}
\arguments{
\item{...}{named slots and their values.}

\item{name}{The name of the slot to set (data, sample_meta or variable_meta
for dataset objects)}

\item{value}{A data.frame}

\item{object, x}{A dataset object}
}
\value{
The returned value depends on the method used
}
\description{
An object for holding raw data and associated meta data
}
\details{
The dataset object is intended to allow raw data and meta data to be kept
together, but without needing to store them e.g. in the same data frame. This
is conventient for statistical methods where the meta data is often required
but not as part of the same matrix as the raw data.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the dataset}

\item{\code{description}}{Brief description of the dataset}

\item{\code{type}}{The type of dataset e.g. single_block}

\item{\code{data}}{A data.frame of data, samples in rows, variables/features
in columns.}

\item{\code{sample_meta}}{A data.frame of sample meta data e.g. group
membership}

\item{\code{variable_meta}}{A data frame of variable meta data}
}}

\examples{
D = dataset()

# get the data from a dataset object
D$data

# get the sample meta data from a dataset object
D$sample_meta

# get the variable meta data from a dataset object
D$variable_meta

}
