% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_node_class.R
\name{struct_node}
\alias{struct_node}
\title{struct_node class}
\usage{
struct_node(name = character(0), description = character(0), ...)
}
\arguments{
\item{name}{the name of the node}

\item{description}{a description of the node}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A base class for representing nodes in a directed acyclic graph (DAG) of struct models.
All node types (data_node, model_node, prediction_node) inherit from this class.
}
\examples{
NODE = struct_node(name = 'Example Node', description = 'A simple example node')
}
