% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_preset_class.R
\name{preset}
\alias{preset}
\title{Struct Preset objects}
\usage{
preset(name, ...)
}
\arguments{
\item{name}{Character string naming the preset}

\item{...}{Additional parameters for the preset}
}
\value{
A list with class 'struct_preset'
}
\description{
A utility function in the \pkg{struct} package for creating preset configurations.
Struct presets are used to store predefined parameter configurations that can be
applied to various objects in the struct framework.
}
\examples{
# Create a preset
p = preset('default', color = 'red', size = 2)
}
