% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\docType{class}
\name{test_method-class}
\alias{test_method-class}
\alias{test_method}
\alias{method.apply,test_method,dataset-method}
\title{Example method}
\usage{
\S4method{method.apply}{test_method,dataset}(M, D)
}
\arguments{
\item{M}{test_method object}

\item{D}{dataset object}
}
\value{
test method object

dataset object
}
\description{
An example model for testing

applies the example method, which adds value_1 to the raw data of a dataset
}
\examples{
M = test_method()
M = test_method(value_1 = 10, value_2 = 20)

D = iris_dataset()
M = test_method(value_1 = 10, value_2 = 20)
M = method.apply(M,D)

}
