% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_class.R
\docType{class}
\name{test_metric-class}
\alias{test_metric-class}
\alias{test_metric}
\alias{calculate,test_metric-method}
\title{Example metric}
\usage{
\S4method{calculate}{test_metric}(obj)
}
\arguments{
\item{obj}{metric object}
}
\value{
test metric object

dataset object
}
\description{
An example metric for testing

calculates a metric, which just returns a value of 3.142
}
\examples{
MET = test_metric()

MET = test_metric()
MET = calculate(MET)

}
