% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_objects.R
\docType{class}
\name{test_model-class}
\alias{test_model-class}
\alias{test_model}
\alias{model.train,test_model,dataset-method}
\alias{model.predict,test_model,dataset-method}
\title{Example model}
\usage{
\S4method{model.train}{test_model,dataset}(M, D)

\S4method{model.predict}{test_model,dataset}(M, D)
}
\arguments{
\item{M}{test_model object}

\item{D}{dataset object}
}
\value{
dataset object

dataset object

dataset object
}
\description{
An example model for testing

trains the example model, which adds value_1 to the raw data of a dataset

predicts using the example model, which adds value_2 to the raw data of a
dataset
}
\examples{
M = test_model()
M = test_model(value_1 = 10, value_2 = 20)

D = iris_dataset()
M = test_model(value_1 = 10, value_2 = 20)
M = model.train(M,D)
D = iris_dataset()
M = test_model(value_1 = 10, value_2 = 20)
M = model.predict(M,D)
}
