% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_class.R
\docType{methods}
\name{*,iterator,model_OR_iterator-method}
\alias{*,iterator,model_OR_iterator-method}
\title{combine an interator with other model or interator objects}
\usage{
\S4method{*}{iterator,model_OR_iterator}(e1, e2)
}
\description{
makes the second input a child of the first input (parent), so that it can be executed multiple times by the parent iterator
}
