% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{train}
\alias{train}
\title{Train a model}
\usage{
train(M, D)
}
\arguments{
\item{model}{a model object}

\item{data}{a dataset object}
}
\description{
Trains a model using the input dataset
}
\seealso{
Other model functions: \code{\link{predicted.name}},
  \code{\link{predicted}}, \code{\link{predict}}
}
