% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{type,struct_class-method}
\alias{type,struct_class-method}
\title{type}
\usage{
\S4method{type}{struct_class}(obj)
}
\arguments{
\item{obj}{a struct object}
}
\value{
the type assigned to the struct objects
}
\description{
get the type of an object
}
\examples{
S = struct_class()
type(S)
}
