% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{type}
\alias{type}
\alias{type<-}
\title{get/set type for an object}
\usage{
type(obj)

type(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid type string}
}
\description{
get/set type for an object
}
\seealso{
Other struct_class functions: \code{\link{description}},
  \code{\link{name}}
}
