% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed_list_class.R
\name{typed_list}
\alias{typed_list}
\alias{show,typed_list-method}
\alias{length,typed_list-method}
\alias{names,typed_list-method}
\alias{names<-,typed_list,character-method}
\alias{[[,typed_list,ANY,ANY-method}
\alias{[[<-,typed_list,ANY,missing-method}
\alias{[,typed_list,ANY,ANY,ANY-method}
\alias{as.list,typed_list-method}
\title{Typed List objects}
\usage{
typed_list(..., .type = NULL, drop_null = FALSE)

\S4method{show}{typed_list}(object)

\S4method{length}{typed_list}(x)

\S4method{names}{typed_list}(x)

\S4method{names}{typed_list,character}(x) <- value

\S4method{[[}{typed_list,ANY,ANY}(x, i, j, ...)

\S4method{[[}{typed_list,ANY,missing}(x, i, j) <- value

\S4method{[}{typed_list,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{as.list}{typed_list}(x, ...)
}
\arguments{
\item{...}{List elements to include}

\item{.type}{Character string describing the type of list}

\item{drop_null}{Logical, whether to remove NULL elements from the list}
}
\value{
A typed list object
}
\description{
A utility class in the \pkg{struct} package for creating typed lists with metadata.
Typed lists are used to store collections of objects with a common type annotation.
}
\examples{
# Create a typed list
tl = typed_list(
    point1 = geom_point(),
    point2 = geom_line(),
    .type = 'Layer',
    drop_null = TRUE
)
}
