% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_class.R
\docType{methods}
\name{value<-,enum-method}
\alias{value<-,enum-method}
\title{set the value of an enum object}
\usage{
\S4method{value}{enum}(obj) <- value
}
\arguments{
\item{obj}{an enum object}

\item{value}{a valid value for the enum}
}
\description{
sets the value for an enum object
}
\examples{
E = enum(list=c('hello','world'))
value(E) = 'hello' # value MUST be in the enum list

}
