% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{variable.meta}
\alias{variable.meta}
\alias{variable.meta<-}
\title{get/set variable meta data for a dataset object}
\usage{
variable.meta(obj)

variable.meta(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of meta data}
}
\description{
get/set variable meta data for a dataset object
}
\seealso{
Other dataset functions: \code{\link{data}},
  \code{\link{sample.meta}}
}
