% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset.variable_meta}
\alias{dataset.variable_meta}
\alias{dataset.variable_meta<-}
\title{get/set variable meta data for a dataset object}
\usage{
dataset.variable_meta(obj)

dataset.variable_meta(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of meta data}
}
\description{
get/set variable meta data for a dataset object
}
\seealso{
Other dataset functions: \code{\link{dataset.data}},
  \code{\link{dataset.sample_meta}}
}
