% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetExperiment_class.R
\name{as.DatasetExperiment,SummarizedExperiment-method}
\alias{as.DatasetExperiment,SummarizedExperiment-method}
\title{Convert a SummarizedExperiment to DatasetExperiment}
\usage{
\S4method{as.DatasetExperiment}{SummarizedExperiment}(obj)
}
\arguments{
\item{obj}{a SummarizedExperiment object}
}
\value{
a DatasetExperiment object
}
\description{
The assay data is transposed, and colData and rowData switched to match. 
struct specific slots such as "name" and "description" are extracted from the 
metaData if available. NB Any additional metadata will be lost during this conversion.
}
