% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R, R/chart_class.R,
%   R/model_class.R, R/metric_class.R, R/iterator_class.R, R/optimiser_class.R,
%   R/preprocess_class.R, R/resampler_class.R
\name{.DollarNames.struct_class}
\alias{.DollarNames.struct_class}
\alias{.DollarNames,struct_class-method}
\alias{.DollarNames.chart}
\alias{.DollarNames,chart-method}
\alias{.DollarNames.model}
\alias{.DollarNames,model-method}
\alias{.DollarNames.metric}
\alias{.DollarNames,metric-method}
\alias{.DollarNames.iterator}
\alias{.DollarNames,iterator-method}
\alias{.DollarNames.optimiser}
\alias{.DollarNames,optimiser-method}
\alias{.DollarNames.preprocess}
\alias{.DollarNames,preprocess-method}
\alias{.DollarNames.resampler}
\alias{.DollarNames,resampler-method}
\title{autocompletion}
\usage{
\method{.DollarNames}{struct_class}(x, pattern = "")

\S4method{.DollarNames}{struct_class}(x, pattern = "")

\method{.DollarNames}{chart}(x, pattern = "")

\S4method{.DollarNames}{chart}(x, pattern = "")

\method{.DollarNames}{model}(x, pattern = "")

\S4method{.DollarNames}{model}(x, pattern = "")

\method{.DollarNames}{metric}(x, pattern = "")

\S4method{.DollarNames}{metric}(x, pattern = "")

\method{.DollarNames}{iterator}(x, pattern = "")

\S4method{.DollarNames}{iterator}(x, pattern = "")

\method{.DollarNames}{optimiser}(x, pattern = "")

\S4method{.DollarNames}{optimiser}(x, pattern = "")

\method{.DollarNames}{preprocess}(x, pattern = "")

\S4method{.DollarNames}{preprocess}(x, pattern = "")

\method{.DollarNames}{resampler}(x, pattern = "")

\S4method{.DollarNames}{resampler}(x, pattern = "")
}
\arguments{
\item{x}{a struct_class object}

\item{pattern}{the text used to compare against the slot names}
}
\value{
A vector of slot names
}
\description{
This function returns slotnames for autocompletion when using $ syntax
}
