% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{chart.names,struct_class-method}
\alias{chart.names,struct_class-method}
\title{chart names}
\usage{
\S4method{chart.names}{struct_class}(obj, ret = "char")
}
\arguments{
\item{obj}{a chart object}

\item{ret}{a string indicating whether to return a list of chart
names (ret="char") or a list of chart objects (ret="obj").
The default is "char".}
}
\value{
list of chart names (default), or chart objects
}
\description{
print a list of chart objects associated with the input object
}
\examples{
S = struct_class()
chart.names(S)
chart.names(S,'char')
chart.names(S,'obj')
}
