% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.names}
\alias{chart.names}
\title{chart names}
\usage{
chart.names(obj, ...)
}
\arguments{
\item{obj}{a object from the *struct* package}

\item{...}{optional inputs}
}
\description{
return a list of valid charts for a struct object
}
\seealso{
Other chart functions: \code{\link{chart.plot}}
}
