% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.plot}
\alias{chart.plot}
\title{plot a chart}
\usage{
chart.plot(obj, name, opt = NULL)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of chart to plot}

\item{opt}{list of options for this chart}
}
\description{
plot a chart
}
\seealso{
Other chart functions: \code{\link{chart.ids}},
  \code{\link{chart.name}}, \code{\link{chart.obj}},
  \code{\link{chart.opt}}, \code{\link{is.chart}}
}
