% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.plot}
\alias{chart.plot}
\title{chart.plot}
\usage{
chart.plot(obj, dobj, ...)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a struct object}

\item{...}{optional inputs}
}
\value{
a plot object
}
\description{
Plots a chart object
}
\examples{
C = example_chart()
chart.plot(C)

}
