% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{chart.plot}
\alias{chart.plot}
\title{plot a chart for an object}
\usage{
chart.plot(obj, dobj, ...)
}
\arguments{
\item{obj}{a chart object}

\item{dobj}{a struct object}

\item{...}{optional inputs}
}
\value{
plot
}
\description{
plot a chart for an object
}
\examples{
\dontrun{

C=chart()
chart.plot(C)
p=chart.plot(C) # can return e.g. ggplot object
}
}
