% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_stato.R
\docType{class}
\name{chart.stato-class}
\alias{chart.stato-class}
\alias{chart.stato}
\title{chart_stato_class}
\description{
A base class in the \pkg{struct} package. Should not be called directly.
}
\details{
Extends the \code{chart} class to include stato functionality.
}
\examples{
obj = chart.stato() # create chart object
}
\seealso{
Refer to \code{\link{chart}} and \code{\link{stato}} for further
info.
}
