% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_class.R
\docType{class}
\name{chart-class}
\alias{chart-class}
\alias{chart}
\alias{chart.plot,chart-method}
\title{chart_class}
\usage{
\S4method{chart.plot}{chart}(obj)
}
\arguments{
\item{obj}{a chart object}
}
\description{
A base class in the \pkg{struct} package. should not be called directly.

plots a chart object
}
\examples{
obj = chart() # create chart object
chart.plot(obj) # plot chart
}
