% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_execute.R, R/model_dag_class.R
\name{dag_execute}
\alias{dag_execute}
\alias{dag_execute,model_dag,logical-method}
\title{Execute a DAG object}
\usage{
dag_execute(dag, verbose = TRUE)

\S4method{dag_execute}{model_dag,logical}(dag, verbose = TRUE)

\S4method{dag_execute}{model_dag,logical}(dag, verbose = TRUE)
}
\arguments{
\item{dag}{a model_dag object}

\item{verbose}{logical, whether to print progress information}
}
\value{
the model_dag object with results included
}
\description{
Executes a directed acyclic graph (DAG) object by traversing the graph
and executing nodes in the correct topological order.
}
\examples{
# Create a simple DAG
D = iris_DatasetExperiment()
M = PCA()

# Create nodes
data_node1 = data_node(name = 'Data', data = D)
model_node1 = model_node(name = 'Model', model = M)

# Create DAG with nodes
dag = model_dag(
    name = 'Simple DAG',
    description = 'A simple example DAG',
    edges = list(
        list(from = 'Data', from_param = 'asis', to = 'Model', to_param = 'input_data')
    ),
    nodes = list(
        'Data' = data_node1,
        'Model' = model_node1
    )
)

# Execute DAG
dag = dag_execute(dag)
# Create a simple DAG
D = iris_DatasetExperiment()
M = example_model()

# Create nodes
data_node1 = data_node(name = 'Data', data = D)
model_node1 = model_node(name = 'Model', model = M)

# Create DAG with nodes
dag = model_dag(
    name = 'Simple DAG',
    description = 'A simple example DAG',
    edges = list(
        list(from = 'Data', to = 'Model')
    ),
    nodes = list(
        'Data' = data_node1,
        'Model' = model_node1
    )
)

# Execute DAG
dag = dag_execute(dag)
}
