% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset_data}
\alias{dataset_data}
\alias{dataset_data<-}
\title{Dataset object data}
\usage{
dataset_data(obj)

dataset_data(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of raw data}
}
\value{
\describe{
\item{\code{dataset_data(obj)}}{returns a data.frame of raw data}
\item{\code{dataset_data(obj)<-}}{sets the raw data of a dataset object}
}
}
\description{
Dataset object data
}
\examples{
# iris dataset
D = iris_dataset()
# original raw data
X = dataset_data(D)
# modify the raw dataset
dataset_data(D) = X + 1
}
