% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset.data}
\alias{dataset.data}
\alias{dataset.data<-}
\title{Dataset object data}
\usage{
dataset.data(obj)

dataset.data(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of raw data}
}
\value{
\describe{
\item{\code{dataset.data(obj)}}{returns a data.frame of raw data}
\item{\code{dataset.data(obj)<-}}{sets the raw data of a dataset object}
}
}
\description{
Dataset object data
}
\examples{
# iris dataset
D = iris_dataset()
# original raw data
X = dataset.data(D)
# modify the raw dataset
dataset.data(D) = X + 1

}
