% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset.data}
\alias{dataset.data}
\alias{dataset.data<-}
\title{get/set data for a dataset object}
\usage{
dataset.data(obj)

dataset.data(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of raw data}
}
\description{
get/set data for a dataset object
}
\seealso{
Other dataset functions: \code{\link{dataset.sample_meta}},
  \code{\link{dataset.variable_meta}}
}
