% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_class.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\alias{dataset.data,dataset-method}
\alias{$,dataset-method}
\alias{dataset.data<-,dataset-method}
\alias{$<-,dataset-method}
\alias{dataset.sample_meta,dataset-method}
\alias{dataset.sample_meta<-,dataset-method}
\alias{dataset.variable_meta,dataset-method}
\alias{dataset.variable_meta<-,dataset-method}
\alias{summary,dataset-method}
\alias{export.xlsx,dataset-method}
\title{Dataset class}
\usage{
\S4method{dataset.data}{dataset}(obj)

\S4method{$}{dataset}(x, name)

\S4method{dataset.data}{dataset}(obj) <- value

\S4method{$}{dataset}(x, name) <- value

\S4method{dataset.sample_meta}{dataset}(obj)

\S4method{dataset.sample_meta}{dataset}(obj) <- value

\S4method{dataset.variable_meta}{dataset}(obj)

\S4method{dataset.variable_meta}{dataset}(obj) <- value

\S4method{summary}{dataset}(object)

\S4method{export.xlsx}{dataset}(object, xlsxfile, transpose = TRUE)
}
\arguments{
\item{obj, object, x}{A dataset object}

\item{name}{The name of the slot to set (data, sample_meta or variable_meta
for dataset objects)}

\item{value}{A data.frame}

\item{xlsxfile}{file name (can include full path) of excel file to write}

\item{transpose}{TRUE (default) or FALSE to transpose the rows/columns when writing an excel sheet}
}
\value{
The returned value depends on the method used
}
\description{
An object for holding raw data and associated meta data
}
\details{
The dataset object is intended to allow raw data and meta data to be kept
together, but without needing to store them e.g. in the same data frame. This
is conventient for statistical methods where the meta data is often required
but not as part of the same matrix as the raw data.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dataset.data}: get the data matrix from a dataset object

\item \code{$}: get data/sample_meta/variable_meta from a dataset object

\item \code{dataset.data<-}: set the data for a dataset object

\item \code{$<-}: set the data/sample_meta/variable_meta for a dataset

\item \code{dataset.sample_meta}: get the sample meta data from a dataset object

\item \code{dataset.sample_meta<-}: set the sample meta data for a dataset

\item \code{dataset.variable_meta}: get the variable meta data from a dataset object

\item \code{dataset.variable_meta<-}: set the variable meta data for a dataset

\item \code{summary}: print a summary of the data set to the terminal

\item \code{export.xlsx}: write the dataset object to an excel file
}}

\section{Slots}{

\describe{
\item{\code{name}}{Name of the dataset}

\item{\code{description}}{Brief description of the dataset}

\item{\code{type}}{The type of dataset e.g. single_block}

\item{\code{data}}{A data.frame of data, samples in rows, variables/features
in columns.}

\item{\code{sample.meta}}{A data.frame of sample meta data e.g. group
membership}

\item{\code{variable.meta}}{A data frame of variable meta data}
}}

\examples{
D = dataset()

# get the data from a dataset object
dataset.data(D) # OR
D$data

# get the sample meta data from a dataset object
dataset.sample_meta(D) # OR
D$sample_meta

# get the variable meta data from a dataset object
dataset.variable_meta(D) # OR
D$variable_meta

}
