% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_class.R
\docType{class}
\name{dataset-class}
\alias{dataset-class}
\alias{dataset}
\alias{dataset.data,dataset-method}
\alias{$,dataset-method}
\alias{dataset.data<-,dataset-method}
\alias{$<-,dataset-method}
\alias{dataset.sample_meta,dataset-method}
\alias{dataset.sample_meta<-,dataset-method}
\alias{dataset.variable_meta,dataset-method}
\alias{dataset.variable_meta<-,dataset-method}
\alias{summary,dataset-method}
\title{dataset class}
\usage{
\S4method{dataset.data}{dataset}(obj)

\S4method{$}{dataset}(x, name)

\S4method{dataset.data}{dataset}(obj) <- value

\S4method{$}{dataset}(x, name) <- value

\S4method{dataset.sample_meta}{dataset}(obj)

\S4method{dataset.sample_meta}{dataset}(obj) <- value

\S4method{dataset.variable_meta}{dataset}(obj)

\S4method{dataset.variable_meta}{dataset}(obj) <- value

\S4method{summary}{dataset}(object)
}
\arguments{
\item{obj}{a dataset object}

\item{x}{a dataset object}

\item{name}{name of the slot to set (data, sample_meta or variable_meta for dataset objects)}

\item{value}{a data.frame}

\item{object}{a dataset object}
}
\description{
an object for holding raw data and associated meta data

get the data matrix from a dataset object

Returns the sample meta data of a dataset object

Sets the sample meta data of a dataset object

Returns the variable metadata of a dataset object

Sets the variable metadata of a dataset object

print a brief summary of the contents of a dataset object. Note that this is different to the output from show(dataset), which summarieses the dataset object, not the contents..
}
\section{Slots}{

\describe{
\item{\code{name}}{(character) name of the dataset}

\item{\code{description}}{(character) brief description of the dataset}

\item{\code{data}}{(data.frame) matrix of data, samples in rows, variables/features in columns.}

\item{\code{sample.meta}}{(data.frame) data frame of sample meta data e.g. group membership}

\item{\code{variable.meta}}{(data.frame) data frame of variable meta data}
}}

\examples{
D = dataset()

D = dataset()
X = dataset.data(D)

D = dataset()
X = D$data
S = D$sample_meta
V = D$variable_meta

D = dataset()
X = dataset.data(D) = iris[,1:4]

D = dataset()
D$data = iris[,1:4]
D$sample_meta = iris[,5,drop=FALSE]
D$variable_meta=data.frame(sample_id = rownames(iris))

D = dataset()
S = dataset.sample_meta(D)

D = dataset()
dataset.sample_meta(D) = iris[,5,drop = FALSE]
D = dataset()
V = dataset.variable_meta(D)
D = dataset()
df = data.frame(sample_id = rownames(iris))
dataset.variable_meta(D)= df

D = dataset(data = iris[,1:4],
                        sample_meta = iris[,5,drop=FALSE])
summary(D)

}
