% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{description}
\alias{description}
\alias{description<-}
\title{get/set description for an object}
\usage{
description(obj)

description(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid string}
}
\description{
get/set description for an object
}
\seealso{
Other struct_class functions: \code{\link{name}},
  \code{\link{type}}
}
\concept{struct_class functions}
