% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_class.R
\docType{class}
\name{entity-class}
\alias{entity-class}
\alias{entity}
\alias{value,entity-method}
\alias{value<-,entity-method}
\title{Entity objects}
\usage{
\S4method{value}{entity}(obj)

\S4method{value}{entity}(obj) <- value
}
\arguments{
\item{obj}{An entity object}

\item{value}{Value of the entity}
}
\description{
A base class in the \pkg{struct} package. Not normally called directly.
}
\details{
An entity object is used to store information about a parameter or output.
The standard 'name','description' and 'type' slots are included, along with
'value' for storing the value of the parameter.

Entity objects are usually defined in the prototype of another object, but
can be extracted using \code{param.obj} and \code{output.obj}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{value}: get the value for an entity

\item \code{value<-}: set the value for an entity
}}

\examples{
# Create a new entity object
E = entity(
    name='example',
    description='this is an example',
    type='numeric',
    value=1
)

# Get/set the value of the entity object
value(E)
value(E) = 10

}
