% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{max_length}
\alias{max_length}
\title{get the max value vector length for an entity}
\usage{
max_length(obj)
}
\arguments{
\item{obj}{an entity object}
}
\value{
max value vector length for an entity
}
\description{
get the max value vector length for an entity
}
