% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/entity_class.R
\name{max_length}
\alias{max_length}
\alias{entity}
\alias{value,entity-method}
\alias{value<-,entity-method}
\alias{max_length,entity-method}
\alias{max_length<-,entity-method}
\title{get the max value vector length for an entity}
\usage{
max_length(obj)

entity(...)

\S4method{value}{entity}(obj)

\S4method{value}{entity}(obj) <- value

\S4method{max_length}{entity}(obj)

\S4method{max_length}{entity}(obj) <- value
}
\arguments{
\item{obj}{An entity object}

\item{...}{named slots and their values.}

\item{value}{Value of the entity}

\item{max_length}{maximum length of value vector (default 1)}
}
\value{
max value vector length for an entity

and entity object
}
\description{
A base class in the \pkg{struct} package. Not normally called directly.
An entity object is used to store information about a parameter or output_
The standard 'name','description' and 'type' slots are included, along with
'value' for storing the value of the parameter and 'max_length' for restricting
the length of 'value' if needed.
}
\details{
Entity objects are usually defined in the prototype of another object, but
can be extracted using \code{param_obj} and \code{output_obj}.
}
\examples{
# Create a new entity object
E = entity(
    name = 'example',
    description = 'this is an example',
    type = 'numeric',
    value = 1
)

# Get/set the value of the entity object
value(E)
value(E) = 10
}
