% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_stato_class.R
\docType{class}
\name{entity.stato-class}
\alias{entity.stato-class}
\alias{entity.stato}
\title{stato_entity_class}
\description{
A base class in the \pkg{struct} package. should not be called directly.
}
