% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/dataset_class.R
\name{export.xlsx}
\alias{export.xlsx}
\alias{export.xlsx,dataset-method}
\title{write a dataset object to file}
\usage{
export.xlsx(object, outfile, transpose)

\S4method{export.xlsx}{dataset}(object, outfile, transpose = TRUE)
}
\arguments{
\item{object}{a dataset object}

\item{outfile}{the filename (including path) to write the data to}

\item{transpose}{TRUE (default) or FALSE to transpose the output data}
}
\value{
an excel file with sheets for data and meta data
}
\description{
Exports a dataset object to an excel file with sheets for data, sample_meta and variable_meta
}
\examples{
\dontrun{
D = iris_dataset() # example dataset
export.xlsx(D,'iris_dataset_xlsx')
}
}
