% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_chart_class.R
\name{ggplot_chart}
\alias{ggplot_chart}
\alias{chart_build,ggplot_chart,DatasetExperiment-method}
\alias{chart_plot,ggplot_chart,DatasetExperiment-method}
\alias{show,ggplot_chart-method}
\title{ggplot Chart objects}
\usage{
ggplot_chart(...)

\S4method{chart_build}{ggplot_chart,DatasetExperiment}(obj, dobj)

\S4method{chart_plot}{ggplot_chart,DatasetExperiment}(obj, dobj, rebuild = NULL)

\S4method{show}{ggplot_chart}(object)
}
\arguments{
\item{...}{Additional parameters passed to chart constructor}
}
\value{
A ggplot_chart object
}
\description{
A base class in the \pkg{struct} package for ggplot2-based charts.
ggplot_chart extends the base chart class to provide common functionality
for all ggplot2 chart objects.
}
\details{
The layers slot can contain any ggplot2 layer including:
- Geoms (geom_point, geom_line, etc.)
- Stats (stat_ellipse, stat_smooth, etc.)
- Scales (scale_color_manual, scale_x_continuous, etc.)
- Themes (theme_minimal, theme_bw, etc.)
- Facets (facet_wrap, facet_grid, etc.)
- Labels (labs, xlab, ylab, etc.)
}
\examples{
# Create a ggplot chart (not typically called directly)
gc = ggplot_chart()
}
