% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is.param}
\alias{is.param}
\title{verify paramater name}
\usage{
is.param(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of parameter as a string}
}
\value{
TRUE if parameter name is valid, FALSE if not
}
\description{
verify that the name of a paramater is valid for an object
}
\seealso{
Other parameter functions: \code{\link{output.name}},
  \code{\link{param.ids}}, \code{\link{param.list}},
  \code{\link{param.name}}, \code{\link{param.obj}},
  \code{\link{param.value}}
}
\concept{parameter functions}
