% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_output}
\alias{is_output}
\title{Verify output}
\usage{
is_output(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of output as a string}
}
\value{
TRUE if output name is valid, FALSE if not
}
\description{
Verify that the name of a output is valid for an object
}
\examples{
M = example_model()
is_output(M,'result_1') # TRUE
is_output(M,'result_0')   # FALSE
}
