% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterator_class.R
\docType{class}
\name{iterator-class}
\alias{iterator-class}
\alias{iterator}
\alias{run,iterator,dataset,metric-method}
\alias{evaluate,iterator,metric-method}
\alias{models,iterator-method}
\alias{models<-,iterator,model_OR_iterator-method}
\alias{result,iterator-method}
\alias{result.name,iterator-method}
\alias{*,iterator,model_OR_iterator-method}
\title{iterator class}
\usage{
\S4method{run}{iterator,dataset,metric}(I, D, MET = NULL)

\S4method{evaluate}{iterator,metric}(I, MET)

\S4method{models}{iterator}(ML)

\S4method{models}{iterator,model_OR_iterator}(ML) <- value

\S4method{models}{iterator,model_OR_iterator}(ML) <- value

\S4method{result}{iterator}(M)

\S4method{result.name}{iterator}(M)

\S4method{*}{iterator,model_OR_iterator}(e1, e2)
}
\arguments{
\item{I}{an iterator object}

\item{D}{a dataset object}

\item{MET}{a metric object}

\item{ML}{a model sequence object}

\item{value}{value}

\item{M}{a model object}

\item{e1}{an iterator object}

\item{e2}{an iterator or a model object}
}
\description{
A class for iterative approaches that involve the training/prediction of a model multiple times.
not intended to be called directly, this class should be inherited to provide functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{run}: run a model/model.seq mutliple times for the input data

\item \code{evaluate}: evaluate the performance of a model/model.seq  using the input metric

\item \code{models}: get the model/model.seq for an iterator object

\item \code{models<-}: set the model/model.seq to be run for multiple iterations

\item \code{models<-}: set the model/model.seq to be run for multiple iterations

\item \code{result}: get prediction output from iterator

\item \code{result.name}: get prediction output name for iterator

\item \code{*}: combine an interator with other model or interator objects
}}

\examples{
I = iterator()
I = iterator() * model()
D = dataset()
MET = metric()
I = iterator() * model()
I = run(I,D,MET)

}
