% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run}
\alias{run}
\alias{evaluate}
\title{run an iterator object}
\usage{
run(I, D, MET)

evaluate(I, MET)
}
\arguments{
\item{I}{an iterator object}

\item{D}{a dataset object}

\item{MET}{a metric object}
}
\description{
run an iterator object

evaluate an iterator object
}
