% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/struct_class.R
\name{libraries}
\alias{libraries}
\alias{libraries,struct_class-method}
\title{Libraries for an object}
\usage{
libraries(obj)

\S4method{libraries}{struct_class}(obj)
}
\arguments{
\item{obj}{a struct object}
}
\value{
a character array of R packages needed by the object
}
\description{
All \code{struct} objects have a "libraries" slot, which is a character array of
libraries required to use the object. The \code{libraries} method gathers
libraries from an object and all \code{struct} objects that it inherits to generate
a complete list.
}
\examples{
M = example_model()
libraries(M)
}
