% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_seq_class.R
\docType{class}
\name{method.seq-class}
\alias{method.seq-class}
\alias{method.seq}
\alias{method.apply,method.seq,dataset-method}
\title{method.seq class}
\usage{
\S4method{method.apply}{method.seq,dataset}(M, D)
}
\description{
A class for (ordered) lists of methods
}
\section{Methods (by generic)}{
\itemize{
\item \code{method.apply}: train the method using input data
}}

