% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{method.steps}
\alias{method.steps}
\alias{method.steps<-}
\title{Get/set methods of a method.seq}
\usage{
method.steps(ML)

method.steps(ML) <- value
}
\arguments{
\item{ML}{a method.seq object}

\item{value}{a list containing only method objects}
}
\description{
Returns the list of method in a method.seq object
}
