% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_class.R
\docType{class}
\name{metric-class}
\alias{metric-class}
\alias{metric}
\alias{calculate,metric-method}
\alias{value,metric-method}
\title{metric class}
\usage{
\S4method{calculate}{metric}(MET, Y, Yhat)

\S4method{value}{metric}(MET)
}
\arguments{
\item{MET}{a metric object}

\item{Y}{= sample meta data e.g. of test data}

\item{Yhat}{= predicted meta data after applying model to test data}
}
\description{
A class for metrics to assess performance of e.g. models, iterators.
Not intended to be called directly, this class should be inherited to provide functionality for method-specific classes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{calculate}: calculate a metric

\item \code{value}: get the caluclated value for a metric
}}

