% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calculate}
\alias{calculate}
\alias{value}
\alias{value<-}
\title{calculate a metric}
\usage{
calculate(MET, Y, Yhat)

value(MET)

value(obj) <- value
}
\arguments{
\item{MET}{a metric object}

\item{Y}{= sample meta data e.g. of test data}

\item{Yhat}{= predicted meta data after applying model to test data}
}
\value{
value the calculated value of a metric
}
\description{
calculate a metric

get the value for a metric
}
