% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calculate}
\alias{calculate}
\alias{value}
\alias{value<-}
\title{calculate a metric}
\usage{
calculate(obj, ...)

value(obj)

value(obj) <- value
}
\arguments{
\item{obj}{a metric object}

\item{...}{additional inputs depending on object}

\item{value}{value}

\item{obj}{metric object}
}
\value{
value the calculated value of a metric
}
\description{
calculate a metric

get the value for a metric

set the valye for a metric
}
