% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/metric_class.R
\name{calculate}
\alias{calculate}
\alias{value}
\alias{value<-}
\alias{max_length<-}
\alias{metric}
\alias{calculate,metric-method}
\alias{value,metric-method}
\alias{value<-,metric-method}
\title{Calculate metric}
\usage{
calculate(obj, ...)

value(obj)

value(obj) <- value

max_length(obj) <- value

metric(...)

\S4method{calculate}{metric}(obj, Y, Yhat)

\S4method{value}{metric}(obj)

\S4method{value}{metric}(obj) <- value
}
\arguments{
\item{obj}{a metric object}

\item{...}{named slots and their values.}

\item{value}{value}

\item{Y}{the true class labels}

\item{Yhat}{the predicted class labels}
}
\value{
value the calculated value of a metric

a metric object
}
\description{
A class for metrics to assess performance of e.g. models, iterators.
Not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\examples{
MET = metric()
calculate(MET)
MET = metric()
M = metric()
calculate(M,Y,Yhat)
MET = metric()
value(MET)
MET = metric()
value(MET) = 10
}
