% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{calculate}
\alias{calculate}
\alias{value}
\alias{value<-}
\title{calculate a metric}
\usage{
calculate(MET, ...)

value(MET)

value(obj) <- value
}
\arguments{
\item{MET}{a metric object}

\item{...}{variable inputs depending on object}

\item{obj}{metric object}

\item{value}{value}
}
\value{
value the calculated value of a metric
}
\description{
calculate a metric

get the value for a metric
}
\concept{metric functions}
