% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_class.R
\docType{class}
\name{model.list-class}
\alias{model.list-class}
\alias{model.list}
\alias{model.train,model.list,dataset-method}
\alias{model.predict,model.list,dataset-method}
\title{model.list class}
\usage{
\S4method{model.train}{model.list,dataset}(M, D)

\S4method{model.predict}{model.list,dataset}(M, D)
}
\description{
A class for (ordered) lists of models
}
\section{Methods (by generic)}{
\itemize{
\item \code{model.train}: train the model using input data

\item \code{model.predict}: apply the model to input data
}}

