% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{model}
\alias{model}
\alias{model_train,model,dataset-method}
\alias{model_predict,model,dataset-method}
\alias{model_apply,model,dataset-method}
\alias{model_reverse,model,dataset-method}
\alias{predicted,model-method}
\alias{predicted_name,model-method}
\alias{predicted_name<-,model,character-method}
\title{model class}
\usage{
model(...)

\S4method{model_train}{model,dataset}(M, D)

\S4method{model_predict}{model,dataset}(M, D)

\S4method{model_apply}{model,dataset}(M, D)

\S4method{model_reverse}{model,dataset}(M, D)

\S4method{predicted}{model}(M)

\S4method{predicted_name}{model}(M)

\S4method{predicted_name}{model,character}(M) <- value
}
\arguments{
\item{...}{named slots and their values.}

\item{M}{a model object}

\item{D}{a dataset object}

\item{value}{value}
}
\value{
trained model object

model object with test set results

trained model object

dataset dataset object with the reverse model applied

the predicted output, as specified by predicted_name

the id of the output returned by predicted()

the modified model object
}
\description{
A class for models that can be trained/applied to datasets e.g. PCA, PLS etc.
Also used for preprocessing steps that require application to test sets.
not intended to be called directly, this class should be inherited to provide
functionality for method-specific classes.
}
\examples{
M = model()
D = dataset()
M = model()
M = model_train(M,D)
D = dataset()
M = model()
M = model_train(M,D)
M = model_predict(M,D)
D = dataset()
M = model()
M = model_apply(M,D)
D = dataset()
M = model()
M = model_train(M,D)
M = model_predict(M,D)
M = model_reverse(M,D)
D = dataset()
M = example_model()
M = model_train(M,D)
M = model_predict(M,D)
p = predicted(M)
M = example_model()
predicted_name(M)
M = example_model()
predicted_name(M) = 'result_2'
}
