% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_list_class.R
\docType{class}
\name{model.seq-class}
\alias{model.seq-class}
\alias{model.seq}
\alias{model.train,model.seq,dataset-method}
\alias{model.predict,model.seq,dataset-method}
\alias{[,model.seq-method}
\alias{[<-,model.seq-method}
\alias{models,model.seq-method}
\alias{models<-,model.seq,list-method}
\alias{length,model.seq-method}
\alias{show,model.seq-method}
\alias{+,model,model.seq-method}
\alias{+,model.seq,model-method}
\alias{+,model,model-method}
\alias{predicted,model.seq-method}
\title{model.seq class}
\usage{
\S4method{model.train}{model.seq,dataset}(M, D)

\S4method{model.predict}{model.seq,dataset}(M, D)

\S4method{[}{model.seq}(x, i)

\S4method{[}{model.seq}(x, i) <- value

\S4method{models}{model.seq}(ML)

\S4method{models}{model.seq,list}(ML) <- value

\S4method{length}{model.seq}(x)

\S4method{show}{model.seq}(object)

\S4method{+}{model,model.seq}(e1, e2)

\S4method{+}{model.seq,model}(e1, e2)

\S4method{+}{model,model}(e1, e2)

\S4method{predicted}{model.seq}(M)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}

\item{x}{a model.seq object}

\item{i}{index}

\item{value}{value}

\item{ML}{a model.seq object}

\item{object}{a model.seq object}

\item{e1}{a model or model.seq object}

\item{e2}{a model or model.seq object}
}
\value{
model sequence

model sequence

model at the given index in the sequence

model sequence with the model at index i replaced

a list of models in the sequence

a model sequence containing the input models

the number of models in the sequence

pritns a summary of the contents of a model sequence

a model sequence with the additional model appended to the front of
the sequence

a model sequence with the additional model appended to the end of the
sequence

a model sequence

the predicted output of the last model in the sequence
}
\description{
A class for (ordered) lists of models
}
\section{Methods (by generic)}{
\itemize{
\item \code{model.train}: train the model using input data

\item \code{model.predict}: apply the model to input data

\item \code{[}: get model by index

\item \code{[<-}: set model by index

\item \code{models}: get a list of models in the sequence

\item \code{models<-}: set the models in the sequence

\item \code{length}: get the number of models in the sequence

\item \code{show}: print a summary of the model.seq to the command line

\item \code{+}: add a model to the (front) of a model sequence

\item \code{+}: add a model to the (end) of a model sequence

\item \code{+}: combine two models into a sequence

\item \code{predicted}: get prediction output from model.seq
}}

\examples{
MS = model.seq()
MS = model() + model()
MS = example_model() + example_model()
MS = model.train(MS,dataset())
D= dataset()
MS = example_model() + example_model()
MS = model.train(MS,D)
MS = model.predict(MS,D)
MS = model() + model()
MS[2]

MS = model() + model()
MS[3] = model()

MS = model() + model()
L = models(MS)

MS = model.seq()
L = list(model(),model())
models(MS) = L

MS = model() + model()
length(MS) # 2

MS = model() + model()
show(MS)

MS = model() + model()
M = model()
MS = M + MS

MS = model() + model()
M = model()
MS = MS + M

MS = model() + model()

D = dataset()
M = example_model()
M = model.train(M,D)
M = model.predict(M,D)
p = predicted(M)
}
