% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dag_class.R
\name{model_dag}
\alias{model_dag}
\title{model_dag class}
\usage{
model_dag(
  name = character(0),
  description = character(0),
  edges = list(),
  nodes = list(),
  results = list(),
  ...
)
}
\arguments{
\item{name}{the name of the DAG}

\item{description}{a description of the DAG}

\item{edges}{a list of edges defining the DAG structure}

\item{...}{additional parameters to pass to struct_class}
}
\description{
A class for representing directed acyclic graphs (DAGs) of struct models.
This class extends struct_class and provides functionality for executing
workflows defined as directed acyclic graphs.
}
\section{\code{edges} slot}{

The "edges" slot is a list that defines the connections between nodes in the DAG.
Each edge should be a list with 'from' and 'to' elements specifying the source
and target nodes.
}

\examples{
DAG = model_dag(name = 'Example DAG', description = 'A simple example DAG')
}
