% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{models}
\alias{models}
\alias{models<-}
\title{Get/set models of a model_seq}
\usage{
models(ML)

models(ML) <- value
}
\arguments{
\item{ML}{a model_seq object}

\item{value}{a list containing only model objects}
}
\value{
\describe{
\item{\code{models(ML)}}{returns a list of models in the model sequence}
\item{\code{models(ML)<-}}{sets the list of models in the model sequence}
}
}
\description{
Returns the list of models in a model_seq object
}
\examples{
# Create a model sequence
ML = model_seq()
models(ML) = list(example_model(), example_model())
models(ML)
}
