% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{name}
\alias{name}
\alias{name<-}
\title{get/set name for an object}
\usage{
name(obj)

name(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid string}
}
\description{
get/set name for an object
}
