% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart_class.R
\name{new_scatter_chart}
\alias{new_scatter_chart}
\alias{chart_build,new_scatter_chart,DatasetExperiment-method}
\alias{show,new_scatter_chart-method}
\title{New Scatter Chart objects}
\usage{
new_scatter_chart(
  xcol = "Petal.Width",
  ycol = "Sepal.Width",
  factor_name = "Species",
  points = layer_entity(layer = "geom_point", value = preset("default"), preset_default =
    "default", presets = list(none = NULL, default = list(mapping = aes(colour =
    .data[["colour"]], shape = .data[["shape"]])))),
  ellipse = layer_entity(layer = "stat_ellipse", value = preset("none"), preset_default =
    "none", presets = list(none = NULL, sample_ellipse = list(mapping = aes(x =
    .data[["x"]], y = .data[["y"]]), colour = "grey", type = "norm"), group_ellipse =
    list(mapping = aes(x = .data[["x"]], y = .data[["y"]], color = .data[["colour"]]),
    type = "norm"))),
  point_labels = layer_entity(layer = "geom_text", value = preset("none"), preset_default
    = "none", presets = list(none = NULL, all_points = list(mapping = aes(x =
    .data[["x"]], y = .data[["y"]], color = .data[["colour"]], label = .data[["label"]]),
    hjust = "outward", vjust = "outward"))),
  ...
)

\S4method{chart_build}{new_scatter_chart,DatasetExperiment}(obj, dobj)

\S4method{show}{new_scatter_chart}(object)
}
\arguments{
\item{xcol}{Character string naming the x-axis column}

\item{ycol}{Character string naming the y-axis column}

\item{factor_name}{Character string naming the factor column for grouping}

\item{points}{Layer entity for points layer}

\item{ellipse}{Layer entity for ellipse layer}

\item{point_labels}{Layer entity for point labels layer}

\item{...}{Additional parameters passed to ggplot_chart constructor}
}
\value{
A new_scatter_chart object
}
\description{
A chart class in the \pkg{struct} package for creating scatter plots using the new ggplot_chart approach.
This chart extends ggplot_chart to provide scatter plot functionality with configurable layers.
}
\examples{
# Create a scatter chart
D = iris_DatasetExperiment()
C = new_scatter_chart(
    xcol = 'Petal.Width',
    ycol = 'Sepal.Width',
    factor_name = 'Species'
)
chart_plot(C, D)
}
