% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.list}
\alias{output.list}
\alias{output.list<-}
\title{output list}
\usage{
output.list(obj)

output.list(obj) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{value}{a named list of outputs for an object}
}
\value{
list of output names

object

list
}
\description{
get a named list of outputs and thier current value for an object
}
\examples{
\dontrun{

output.list(M)
}
\dontrun{

output.list(M) = liist('value_1'=10)
}
}
