% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.list}
\alias{output.list}
\alias{output.list<-}
\title{output list}
\usage{
output.list(obj)

output.list(obj) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{value}{a named list}
}
\value{
list of output names

struct object
}
\description{
get/set a named list of outputs and their current value for an object
}
\examples{
M = example_model()
L = output.list(M)

M = example_model()
output.list(M) = list('result_1' = dataset(),'result_2' = dataset())

}
