% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.name}
\alias{output.name}
\title{output name}
\usage{
output.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{id of output}
}
\value{
name of output
}
\description{
return a the name for a output, if available
}
\examples{
M = test_model()
output.name(M,'result_1')

}
