% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.name}
\alias{output.name}
\title{output names}
\usage{
output.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
list of parameter names
}
\description{
return a list of valid parameter names for an object
}
\seealso{
Other parameter functions: \code{\link{is.param}},
  \code{\link{param.ids}}, \code{\link{param.list}},
  \code{\link{param.name}}, \code{\link{param.obj}},
  \code{\link{param.value}}
}
