% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.value}
\alias{output.value}
\alias{output.value<-}
\title{output values}
\usage{
output.value(obj, name)

output.value(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{of an output}

\item{value}{value}
}
\value{
value of output

struct object
}
\description{
get/set the values for an output.
}
\examples{
M = example_model()
output.value(M,'result_1')

M = example_model()
output.value(M,'result_1') = dataset()

}
