% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output_ids}
\alias{output_ids}
\title{Output identifiers}
\usage{
output_ids(obj)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
list of output ids
}
\description{
return a list of valid output ids for an object
}
\examples{
M = example_model()
output_ids(M)
}
