% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output_list}
\alias{output_list}
\alias{output_list<-}
\title{output list}
\usage{
output_list(obj)

output_list(obj) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{value}{a named list}
}
\value{
list of output names

struct object
}
\description{
get/set a named list of outputs and their current value for an object
}
\examples{
M = example_model()
L = output_list(M)
M = example_model()
output_list(M) = list('result_1' = DatasetExperiment(),'result_2' = DatasetExperiment())
}
