% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{output.obj}
\alias{output.obj}
\alias{output.obj<-}
\title{get/set outputs for an object}
\usage{
output.obj(obj, name)

output.obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of output as a string}

\item{value}{a valid value for the parameter being set}
}
\description{
get/set outputs for an object
}
