% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_class.R
\docType{class}
\name{outputs_class-class}
\alias{outputs_class-class}
\alias{outputs_class}
\title{outputs_class}
\description{
A base class in the \pkg{struct} package. Provides several fundamental methods for getting/setting outputs etc and should not be called directly.
}
