% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.ids}
\alias{param.ids}
\title{parameter identifiers}
\usage{
param.ids(obj)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
list of parameter ids
}
\description{
return a list of valid parameter ids for an object
}
\examples{
M = example_model()
param.ids(M)

}
