% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.list}
\alias{param.list}
\alias{param.list<-}
\title{parameter list}
\usage{
param.list(obj)

param.list(obj) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{value}{a named list, where the name-value pairs should match valid parameters for the object}
}
\value{
list of parameter names
}
\description{
get/set a named list of parameters and thier current value for an object
}
