% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.name}
\alias{param.name}
\title{parameter name}
\usage{
param.name(obj, name)
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}
}
\value{
name of parameter
}
\description{
return a the name for a paramater, if available
}
\seealso{
Other parameter functions: \code{\link{is.param}},
  \code{\link{output.name}}, \code{\link{param.ids}},
  \code{\link{param.list}}, \code{\link{param.obj}},
  \code{\link{param.value}}
}
