% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.obj}
\alias{param.obj}
\alias{param.obj<-}
\title{get/set parameters for an object}
\usage{
param.obj(obj, name)

param.obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of parameter as a string}

\item{value}{a valid value for the parameter being set}
}
\description{
get/set parameters for an object
}
\seealso{
Other parameter functions: \code{\link{is.param}},
  \code{\link{output.name}}, \code{\link{param.ids}},
  \code{\link{param.list}}, \code{\link{param.name}},
  \code{\link{param.value}}
}
\concept{parameter functions}
