% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.obj}
\alias{param.obj}
\alias{param.obj<-}
\title{Get/set parameters for an object}
\usage{
param.obj(obj, name)

param.obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of parameter as a string}

\item{value}{a valid value for the parameter being set}
}
\value{
\describe{
\item{\code{param.obj(M,name)}}{returns the named parameter as an object}
\item{\code{param.obj(M,name)<-}}{sets the named parameter of an object}
}
}
\description{
Gets or sets the object of a parameter e.g. to an entity() object.
}
\examples{
# get the parameter as an object
M = test_model()
obj = param.obj(M, 'value_1')

# set a parameter as an object
param.obj(M, 'value_1') = entity(value = 15)


}
