% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param.value}
\alias{param.value}
\alias{param.value<-}
\title{parameter values}
\usage{
param.value(obj, name)

param.value(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{of a parameter}

\item{value}{value}
}
\value{
value of parameter

struct object
}
\description{
get/set the values for a parameter.
}
\examples{
M = example_model()
param.value(M,'value_1')

M = example_model()
param.value(M,'value_1') = 0.95

}
