% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{param_obj}
\alias{param_obj}
\alias{param_obj<-}
\title{Parameter objects}
\usage{
param_obj(obj, name)

param_obj(obj, name) <- value
}
\arguments{
\item{obj}{a model or iterator object from the *struct* class}

\item{name}{name of parameter as a string}

\item{value}{a valid value for the parameter being set}
}
\value{
\describe{
\item{\code{param_obj(M,name)}}{returns the named parameter as an object}
\item{\code{param_obj(M,name)<-}}{sets the named parameter of an object}
}
}
\description{
Gets or sets the object of a parameter e.g. to an entity() object.
}
\examples{
# get the parameter as an object
M = example_model()
obj = param_obj(M, 'value_0')

# set a parameter as an object
param_obj(M, 'value_0') = entity(value = 15,type = 'numeric')

}
