% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict}
\alias{predict}
\title{Apply a (trained) model to a dataset D to get a prediction}
\usage{
predict(M, D)
}
\arguments{
\item{model}{a model object}

\item{data}{a dataset object}
}
\description{
Apply a model using the input dataset. Requires that the model is trained first.
}
\seealso{
Other model functions: \code{\link{predicted.name}},
  \code{\link{predicted}}, \code{\link{train}}
}
