% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_predict}
\alias{model_predict}
\title{Model prediction}
\usage{
model_predict(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Apply a model using the input dataset_ Assumes the model is trained
first.
}
\examples{
M = example_model()
M = model_predict(M,iris_dataset())

}
