% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model.predict}
\alias{model.predict}
\title{Model prediction}
\usage{
model.predict(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Apply a model using the input dataset. Assumes the model is trained
first.
}
\examples{
M = test_model()
M = model.predict(M,iris_dataset())

}
