% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predicted.name}
\alias{predicted.name}
\title{get/set output name as prediction output for a model}
\usage{
predicted.name(M)
}
\arguments{
\item{M}{a model object}

\item{value}{name of an output for this model}
}
\description{
get/set the prediction output for a model. This determines which outputs from this model are supplied as inputs to the next model when used in a model.list
}
\seealso{
Other model functions: \code{\link{predicted}},
  \code{\link{predict}}, \code{\link{train}}
}
