% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predicted}
\alias{predicted}
\title{Prediction output}
\usage{
predicted(M)
}
\arguments{
\item{M}{a model object}
}
\value{
The value returned varies depending on the output_
}
\description{
returns the prediction output for a model_ This is supplied as input to the
next model when used in a model_seq
}
\examples{
M = example_model()
M = model_train(M, iris_dataset())
M = model_predict(M, iris_dataset())
predicted(M)
}
