% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{result}
\alias{result}
\title{results output for an iterator}
\usage{
result(M)
}
\arguments{
\item{M}{an iterator object}
}
\description{
returns the results of an iterator. This is used to control model flow.
}
\seealso{
Other iterator functions: \code{\link{result.name}},
  \code{\link{run}}
}
\concept{iterator functions}
