% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\docType{methods}
\name{show,struct_class-method}
\alias{show,struct_class-method}
\title{object summary}
\usage{
\S4method{show}{struct_class}(object)
}
\arguments{
\item{object}{a struct_class object}
}
\value{
brief description of the struct object
}
\description{
print a brief summary of a struct object
}
\examples{
S = struct_class()
show(S)
}
