% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{stato.id}
\alias{stato.id}
\alias{stato.name}
\alias{stato.definition}
\alias{stato.summary}
\title{get the stato.id for an object}
\usage{
stato.id(obj)

stato.name(obj)

stato.definition(obj)

stato.summary(obj)
}
\arguments{
\item{obj}{stato_class object}
}
\value{
id the stato id

name the stato name

def the stato description
}
\description{
get the stato.id for an object

get the stato name for an object

get the stato name for an object

display a stato summary of the object
}
\examples{
M = example_model()
stato.id(M)
stato.name(M)
stato.definition(M)
stato.summary(M)

}
