% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{stato.id}
\alias{stato.id}
\alias{stato.name}
\alias{stato.definition}
\alias{stato.summary}
\alias{stato.pdf}
\title{get the stato.id for an object}
\usage{
stato.id(obj)

stato.name(obj)

stato.definition(obj)

stato.summary(obj)

stato.pdf(obj, outpath, filename, outformat)
}
\arguments{
\item{obj}{stato_class object}

\item{outpath}{folder path}

\item{filename}{file name}

\item{outformat}{output format ("pdf" or "html")}
}
\value{
id the stato id

name the stato name

def the stato description
}
\description{
get the stato.id for an object

get the stato name for an object

get the stato name for an object

display a stato summary of the object

get a pdf of a stato object
}
