% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct-package.R
\docType{package}
\name{struct-package}
\alias{struct}
\alias{struct-package}
\title{struct: Strict modification of objects}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: antoine fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
