% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct-package.R
\docType{package}
\name{struct-package}
\alias{struct}
\alias{struct-package}
\title{struct: Strict modification of objects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/struct}
  \item Report bugs at \url{https://github.com/cynkra/struct/issues}
}

}
\author{
\strong{Maintainer}: antoine fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
