% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\docType{package}
\name{struct}
\alias{struct}
\alias{struct-package}
\title{struct: Statistics in R Using Class Templates}
\description{
This package defines classes for developing statistical workflows.
The idea is that these classes can be extended using other packages, making it easier
to combine methods from different packages into a robust workflow.
}
\section{Classes}{

The classes include a \strong{dataset} object to standardise input of data into methods,
as well as a \strong{model} class for methods such as PLS, and an \strong{iterator} class for methods
such as permutation testing and cross validation.
}

