% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\name{struct_class}
\alias{struct_class}
\title{struct_class}
\usage{
struct_class(...)
}
\arguments{
\item{...}{named slots and their values.}
}
\value{
an struct object
}
\description{
A base class in the \pkg{struct} package. Provides several fundamental
methods for getting/setting parameters etc and should not be called directly.
}
\section{Slots}{

\describe{
\item{\code{name}}{name of object}

\item{\code{description}}{description of object}

\item{\code{type}}{type of object}

\item{\code{libraries}}{R packages needed by this object}
}}

\examples{
S = struct_class()
}
