% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetExperiment_class.R
\name{DatasetExperiment}
\alias{DatasetExperiment}
\alias{$,DatasetExperiment-method}
\alias{$<-,DatasetExperiment-method}
\title{DatasetExperiment class}
\usage{
DatasetExperiment(...)

\S4method{$}{DatasetExperiment}(x, name)

\S4method{$}{DatasetExperiment}(x, name) <- value
}
\arguments{
\item{...}{slots and values to populate a new DatasetExperiment with}

\item{x}{A DatasetExperiment object}

\item{name}{DatasetExperiment slot to get/set}

\item{value}{the value to assign to the named slot}
}
\value{
DatasetExperiment
}
\description{
An object for holding raw data and associated meta data
}
\details{
The DatasetExperiment object is an extension of the SummarizedExperiment object 
from the SummarizedExperiment package (found on Bioconductor). 
It incorporates the basic functionality of struct objects, containing fields such as 
Description, Name and Type with features of SummarizedExperiment such as subsetting.

There are some important differences between DatasetExperiment and SummarizedExperiment:
\itemize{
\item In DatasetExperiment data is stored as Samples (rows) x Features (columns)
\item DatasetExperiment currently only supports a single assay
\item length(DatasetExperiment) returns the number of samples
}
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the dataset}

\item{\code{description}}{Brief description of the dataset}

\item{\code{type}}{The type of dataset e.g. single_block}
}}

