% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model.train}
\alias{model.train}
\title{Train a model}
\usage{
model.train(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Trains a model using the input dataset
}
\examples{
M = example_model()
M = model.train(M,iris_dataset())

}
