% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model.train}
\alias{model.train}
\title{Train a model}
\usage{
model.train(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\description{
Trains a model using the input dataset
}
\seealso{
Other model functions: \code{\link{model.predict}},
  \code{\link{predicted.name}}, \code{\link{predicted}}
}
