% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_train}
\alias{model_train}
\title{Train a model}
\usage{
model_train(M, D)
}
\arguments{
\item{M}{a model object}

\item{D}{a dataset object}
}
\value{
Returns a modified model object
}
\description{
Trains a model using the input dataset
}
\examples{
M = example_model()
M = model_train(M,iris_DatasetExperiment())
}
