% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct_class.R
\docType{methods}
\name{type,struct_class-method}
\alias{type,struct_class-method}
\title{type}
\usage{
\S4method{type}{struct_class}(obj)
}
\arguments{
\item{obj}{a struct object}

\item{value}{value}
}
\description{
get the type of an object
}
