% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/struct_class.R
\docType{methods}
\name{type}
\alias{type}
\alias{type<-}
\alias{type<-,struct_class-method}
\title{get/set type for an object}
\usage{
type(obj)

type(obj) <- value

\S4method{type}{struct_class}(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid type string}
}
\description{
get/set type for an object
}
\section{Methods (by class)}{
\itemize{
\item \code{struct_class}: set the type of an object
}}

\seealso{
Other struct_class functions: \code{\link{description}},
  \code{\link{name}}
}
