% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/struct_class.R
\name{type}
\alias{type}
\alias{type<-}
\alias{type<-,struct_class-method}
\title{Object Type}
\usage{
type(obj)

type(obj) <- value

\S4method{type}{struct_class}(obj) <- value
}
\arguments{
\item{obj}{an object from the \pkg{struct} class}

\item{value}{a valid type string}
}
\value{
\describe{
\item{\code{type(obj)}}{returns the type of an object as a string}
\item{\code{type(obj)<-}}{sets the type of an object and returns the modified
object}
}

modified struct object
}
\description{
Get/set the type of a object. Can be used to specify e.g. where a model
is for classification or for regression, or a dataset is single or multi
block for example.
}
\section{Methods (by class)}{
\itemize{
\item \code{struct_class}: set the type of an object
}}

\examples{
M = example_model()
type(M)
type(M) = 'example'
S = struct_class()
type(S) = 'example'
}
