% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dataset_variable_meta}
\alias{dataset_variable_meta}
\alias{dataset_variable_meta<-}
\title{Variable meta data}
\usage{
dataset_variable_meta(obj)

dataset_variable_meta(obj) <- value
}
\arguments{
\item{obj}{a type object from the *struct* class}

\item{value}{a data.frame of meta data}
}
\value{
\describe{
\item{\code{dataset_variable_meta(obj)}}{returns a data.frame of meta data}
\item{\code{dataset_variable_meta(obj)<-}}{sets the meta data of a dataset
object}
}
}
\description{
get/set variable meta data for a dataset object
}
\examples{
# iris dataset
D = iris_dataset()
X = dataset_variable_meta(D)
dataset_variable_meta(D) = X
}
